Split MarkdownHTML.swift into per-concern extension files
The 4,087-line file becomes a 469-line core (config, tokens, the render
pipeline) plus nine nonisolated extensions: RTL, Footnotes, Math, KaTeX,
HostBridge, Highlight, Mermaid, Stylesheet, and shared Utils. About 60%
of the original file was JS/CSS string literals; the three largest
(hostBridgeScript, stylesheet, mermaidInitWiring) now live alone.

Code moves only, except that members called across the new file
boundaries lose `private` (extensions in other files cannot see them).
The extensions are marked nonisolated because the targets default to
MainActor isolation and rendering runs off the main actor.

New files are added to the quick-look target's membership exceptions and
symlinked into the SwiftPM test package.
