+++ md-preview/Rendering/MarkdownHTML.swift
@@ -33,7 +33,7 @@ nonisolated enum MarkdownHTML {
     /// parser paint the document text before it grinds through the vendor
     /// JS — `.inline`'s answer to `.lazy`'s deferred fetch. `.lazy` emissions
     /// keep everything in `head`, byte-identical to the pre-split output.
-    private struct VendorEmission {
+    struct VendorEmission {
         var head: String = ""
         var body: String = ""
     }
@@ -465,3623 +465,5 @@ nonisolated enum MarkdownHTML {
         return result
     }
 
-    // MARK: - RTL Direction
 
-    // Matches opening block-level tags whose direction controls alignment and
-    // logical-edge styling such as blockquote borders.
-    private static let rtlTagRegex: NSRegularExpression = {
-        // swiftlint:disable:next force_try
-        try! NSRegularExpression(
-            pattern: #"<(blockquote|p|li|h[1-6])(\s[^>]*)?>"#,
-            options: [.caseInsensitive]
-        )
-    }()
-
-    private static let htmlTagRegex: NSRegularExpression = {
-        // swiftlint:disable:next force_try
-        try! NSRegularExpression(pattern: #"<[^>]+>"#)
-    }()
-
-    // RTL Unicode ranges: Hebrew, Arabic (+ supplements), Syriac, Thaana, N'Ko, Samaritan, Mandaic
-    private static let rtlRanges: [ClosedRange<UInt32>] = [
-        0x0590...0x05FF, 0x0600...0x06FF, 0x0700...0x074F, 0x0750...0x077F,
-        0x0780...0x07BF, 0x07C0...0x07FF, 0x0800...0x083F, 0x0840...0x085F,
-        0x08A0...0x08FF, 0xFB50...0xFDFF, 0xFE70...0xFEFF
-    ]
-
-    /// Over-inclusive by design: CommonMark decodes numeric character
-    /// references before formatting, so any `&#` might become RTL text.
-    /// False positives only take the existing slow path; false negatives
-    /// would skip direction inference.
-    private static func sourceMayNeedRTLDirection(_ source: String) -> Bool {
-        var previousWasAmpersand = false
-        for scalar in source.unicodeScalars {
-            if scalar.value >= 0x0590,
-               scalar.value <= 0xFEFF,
-               isRTL(scalar) {
-                return true
-            }
-            if previousWasAmpersand, scalar.value == 0x23 {
-                return true
-            }
-            previousWasAmpersand = scalar.value == 0x26
-        }
-        return false
-    }
-
-    private static func injectRTLDirection(in html: String) -> String {
-        let nsHtml = html as NSString
-        let matches = rtlTagRegex.matches(in: html, range: NSRange(location: 0, length: nsHtml.length))
-        guard !matches.isEmpty else { return html }
-
-        var result = ""
-        result.reserveCapacity(html.count + matches.count * 12)
-        var cursor = 0
-
-        for match in matches {
-            result += nsHtml.substring(with: NSRange(location: cursor, length: match.range.location - cursor))
-            let tag = nsHtml.substring(with: match.range(at: 1))
-            let attrs = match.range(at: 2).location != NSNotFound ? nsHtml.substring(with: match.range(at: 2)) : ""
-
-            if attrs.lowercased().contains("dir=") {
-                result += nsHtml.substring(with: match.range)
-            } else {
-                let contentStart = match.range.location + match.range.length
-                let maxLookahead = min(300, nsHtml.length - contentStart)
-                let contentPreview = nsHtml.substring(with: NSRange(location: contentStart, length: maxLookahead))
-                let plainText = stripHTMLTags(contentPreview)
-
-                if let first = firstStrongCharacter(in: plainText), isRTL(first) {
-                    result += "<\(tag)\(attrs) dir=\"rtl\">"
-                } else {
-                    result += nsHtml.substring(with: match.range)
-                }
-            }
-            cursor = match.range.location + match.range.length
-        }
-        result += nsHtml.substring(from: cursor)
-        return result
-    }
-
-    private static func stripHTMLTags(_ html: String) -> String {
-        let nsStr = html as NSString
-        return htmlTagRegex.stringByReplacingMatches(
-            in: html, range: NSRange(location: 0, length: nsStr.length), withTemplate: ""
-        )
-    }
-
-    private static func firstStrongCharacter(in text: String) -> Character? {
-        text.first { char in
-            guard let scalar = char.unicodeScalars.first else { return false }
-            switch scalar.properties.generalCategory {
-            case .uppercaseLetter, .lowercaseLetter, .titlecaseLetter,
-                 .modifierLetter, .otherLetter, .nonspacingMark, .spacingMark, .enclosingMark:
-                return true
-            default:
-                return false
-            }
-        }
-    }
-
-    private static func isRTL(_ char: Character) -> Bool {
-        guard let scalar = char.unicodeScalars.first else { return false }
-        return isRTL(scalar)
-    }
-
-    private static func isRTL(_ scalar: Unicode.Scalar) -> Bool {
-        return rtlRanges.contains { $0.contains(scalar.value) }
-    }
-
-    // MARK: - Footnotes
-
-    private struct FootnoteExtraction {
-        let markdown: String
-        let definitions: [FootnoteDefinition]
-        let references: [FootnoteReference]
-    }
-
-    private struct FootnoteDefinition {
-        let key: String
-        let label: String
-        let content: String
-        let number: Int
-        let sourceLine: Int
-    }
-
-    private struct FootnoteReference {
-        let number: Int
-        let ordinal: Int
-    }
-
-    private struct FootnoteDefinitionRenderResult {
-        let html: String
-        let containsMath: Bool
-        let containsMermaid: Bool
-    }
-
-    private static let footnoteDefinitionRegex: NSRegularExpression = {
-        // swiftlint:disable:next force_try
-        try! NSRegularExpression(pattern: #"^[ \t]{0,3}\[\^([^\]\n]+)\]:[ \t]*(.*)$"#)
-    }()
-
-    private static let footnoteReferenceRegex: NSRegularExpression = {
-        // swiftlint:disable:next force_try
-        try! NSRegularExpression(pattern: #"\[\^([^\]\n]+)\]"#)
-    }()
-
-    private static func extractFootnotes(from markdown: String) -> FootnoteExtraction {
-        guard markdown.contains("[^") else {
-            return FootnoteExtraction(markdown: markdown, definitions: [], references: [])
-        }
-
-        let split = splitFootnoteDefinitions(from: markdown)
-        var protected: [String] = []
-
-        let afterFences = replaceFullMatches(of: codeFenceRegex, in: split.markdown) { full in
-            protected.append(full)
-            return "MdPreviewFootnoteProtect\(protected.count - 1)Token"
-        }
-        let afterInlineCode = replaceFullMatches(of: inlineCodeRegex, in: afterFences) { full in
-            protected.append(full)
-            return "MdPreviewFootnoteProtect\(protected.count - 1)Token"
-        }
-
-        var orderedDefinitions: [FootnoteDefinition] = []
-        var referenceOrdinalsByNumber: [Int: Int] = [:]
-        var references: [FootnoteReference] = []
-
-        let replacedReferences = replaceFootnoteReferenceMatches(in: afterInlineCode) { label, full in
-            let key = normalizeFootnoteKey(label)
-            guard let stored = split.definitions[key] else { return full }
-
-            let definition: FootnoteDefinition
-            if let existing = orderedDefinitions.first(where: { $0.key == key }) {
-                definition = existing
-            } else {
-                definition = FootnoteDefinition(
-                    key: key,
-                    label: stored.label,
-                    content: stored.content,
-                    number: orderedDefinitions.count + 1,
-                    sourceLine: stored.sourceLine
-                )
-                orderedDefinitions.append(definition)
-            }
-
-            let ordinal = (referenceOrdinalsByNumber[definition.number] ?? 0) + 1
-            referenceOrdinalsByNumber[definition.number] = ordinal
-            let token = "MdPreviewFootnoteRef\(references.count)Token"
-            references.append(FootnoteReference(number: definition.number, ordinal: ordinal))
-            return token
-        }
-
-        let restored = restoreIndexedTokens(
-            in: replacedReferences,
-            prefix: "MdPreviewFootnoteProtect",
-            suffix: "Token",
-            replacements: protected
-        )
-
-        return FootnoteExtraction(
-            markdown: restored,
-            definitions: orderedDefinitions,
-            references: references
-        )
-    }
-
-    private static func splitFootnoteDefinitions(from markdown: String) -> (
-        markdown: String,
-        definitions: [String: (label: String, content: String, sourceLine: Int)]
-    ) {
-        let lines = markdown.components(separatedBy: "\n")
-        var output: [String] = []
-        var definitions: [String: (label: String, content: String, sourceLine: Int)] = [:]
-        var index = 0
-
-        while index < lines.count {
-            let line = lines[index]
-            if let match = firstMatch(of: footnoteDefinitionRegex, in: line) {
-                let definitionStartIndex = index
-                let nsLine = line as NSString
-                let label = nsLine.substring(with: match.range(at: 1))
-                var contentLines = [nsLine.substring(with: match.range(at: 2))]
-                index += 1
-
-                while index < lines.count {
-                    let continuation = lines[index]
-                    if continuation.trimmingCharacters(in: .whitespaces).isEmpty {
-                        if index + 1 < lines.count, isIndentedFootnoteContinuation(lines[index + 1]) {
-                            contentLines.append("")
-                            index += 1
-                            continue
-                        }
-                        break
-                    }
-                    guard isIndentedFootnoteContinuation(continuation) else { break }
-                    contentLines.append(stripFootnoteContinuationIndent(from: continuation))
-                    index += 1
-                }
-
-                definitions[normalizeFootnoteKey(label)] = (
-                    label: label,
-                    content: contentLines.joined(separator: "\n"),
-                    sourceLine: definitionStartIndex + 1
-                )
-                // Keep one blank placeholder for every removed source line.
-                // Swift Markdown ignores these lines, while ranges for all
-                // following blocks continue to match the original document.
-                output.append(contentsOf: repeatElement(
-                    "",
-                    count: index - definitionStartIndex
-                ))
-            } else {
-                output.append(line)
-                index += 1
-            }
-        }
-
-        return (output.joined(separator: "\n"), definitions)
-    }
-
-    private static func firstMatch(of regex: NSRegularExpression,
-                                   in source: String) -> NSTextCheckingResult? {
-        let nsSource = source as NSString
-        return regex.firstMatch(
-            in: source,
-            range: NSRange(location: 0, length: nsSource.length)
-        )
-    }
-
-    private static func isIndentedFootnoteContinuation(_ line: String) -> Bool {
-        if line.hasPrefix("\t") { return true }
-        return line.count >= 4 && line.prefix(4).allSatisfy { $0 == " " }
-    }
-
-    private static func stripFootnoteContinuationIndent(from line: String) -> String {
-        if line.hasPrefix("\t") {
-            return String(line.dropFirst())
-        }
-        if line.count >= 4 && line.prefix(4).allSatisfy({ $0 == " " }) {
-            return String(line.dropFirst(4))
-        }
-        return line
-    }
-
-    private static func normalizeFootnoteKey(_ label: String) -> String {
-        label.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
-    }
-
-    private static func replaceFootnoteReferenceMatches(in source: String,
-                                                        transform: (String, String) -> String) -> String {
-        let nsSource = source as NSString
-        let matches = footnoteReferenceRegex.matches(
-            in: source,
-            range: NSRange(location: 0, length: nsSource.length)
-        )
-        guard !matches.isEmpty else { return source }
-
-        var result = ""
-        result.reserveCapacity(source.count)
-        var cursor = 0
-        for match in matches {
-            result += nsSource.substring(with: NSRange(
-                location: cursor,
-                length: match.range.location - cursor
-            ))
-            let full = nsSource.substring(with: match.range)
-            let label = nsSource.substring(with: match.range(at: 1))
-            result += transform(label, full)
-            cursor = match.range.location + match.range.length
-        }
-        result += nsSource.substring(from: cursor)
-        return result
-    }
-
-    private static func renderFootnoteReferences(in html: String,
-                                                 with footnotes: FootnoteExtraction) -> String {
-        guard !footnotes.references.isEmpty else { return html }
-        let replacements = footnotes.references.map { reference in
-            let refID = footnoteReferenceID(number: reference.number, ordinal: reference.ordinal)
-            let footnoteID = footnoteDefinitionID(number: reference.number)
-            let accessibilityLabel = htmlEscape(String(
-                format: NSLocalizedString("Footnote %d", comment: "Footnote accessibility label"),
-                reference.number
-            ))
-            let replacement = """
-            <sup class="footnote-ref"><a id="\(refID)" href="#\(footnoteID)" aria-label="\(accessibilityLabel)">\(reference.number)</a></sup>
-            """
-            return replacement
-        }
-        return restoreIndexedTokens(
-            in: html,
-            prefix: "MdPreviewFootnoteRef",
-            suffix: "Token",
-            replacements: replacements
-        )
-    }
-
-    private static func renderFootnoteDefinitions(
-        _ footnotes: FootnoteExtraction,
-        sourceLineOffset: Int
-    ) -> FootnoteDefinitionRenderResult {
-        guard !footnotes.definitions.isEmpty else {
-            return FootnoteDefinitionRenderResult(
-                html: "",
-                containsMath: false,
-                containsMermaid: false
-            )
-        }
-
-        var containsMath = false
-        var containsMermaid = false
-        let referencesByNumber = Dictionary(grouping: footnotes.references, by: { $0.number })
-        let items = footnotes.definitions.map { definition -> String in
-            let renderedContent = renderFootnoteDefinitionContent(
-                definition.content,
-                sourceLineOffset: sourceLineOffset + definition.sourceLine - 1
-            )
-            containsMath = containsMath || renderedContent.containsMath
-            containsMermaid = containsMermaid || renderedContent.containsMermaid
-            let backrefs = (referencesByNumber[definition.number] ?? []).map { reference in
-                let accessibilityLabel = htmlEscape(String(
-                    format: NSLocalizedString(
-                        "Back to reference %d",
-                        comment: "Footnote back-reference accessibility label"
-                    ),
-                    reference.number
-                ))
-                return """
-                <a href="#\(footnoteReferenceID(number: reference.number, ordinal: reference.ordinal))" class="footnote-backref" aria-label="\(accessibilityLabel)">&#8617;</a>
-                """
-            }.joined(separator: " ")
-            let contentHTML = appendFootnoteBackrefs(backrefs, to: renderedContent.html)
-
-            return """
-            <li id="\(footnoteDefinitionID(number: definition.number))">
-            \(contentHTML)
-            </li>
-            """
-        }.joined(separator: "\n")
-
-        return FootnoteDefinitionRenderResult(
-            html: """
-
-            <section class="footnotes" role="doc-endnotes">
-            <hr />
-            <ol>
-            \(items)
-            </ol>
-            </section>
-            """,
-            containsMath: containsMath,
-            containsMermaid: containsMermaid
-        )
-    }
-
-    private static func appendFootnoteBackrefs(_ backrefs: String, to html: String) -> String {
-        guard !backrefs.isEmpty else { return html }
-        let inlineBackrefs = "<span class=\"footnote-backrefs\">\(backrefs)</span>"
-        if let range = html.range(of: "</p>", options: .backwards) {
-            var updated = html
-            updated.replaceSubrange(range, with: " \(inlineBackrefs)</p>")
-            return updated
-        }
-        return html + inlineBackrefs
-    }
-
-    private static func renderFootnoteDefinitionContent(
-        _ markdown: String,
-        sourceLineOffset: Int
-    ) -> FootnoteDefinitionRenderResult {
-        let math = extractMath(from: markdown.trimmingCharacters(in: .whitespacesAndNewlines))
-        let formatted = EscapingHTMLFormatter.format(
-            math.processedMarkdown,
-            sourceLineOffset: sourceLineOffset
-        )
-        let mermaidResult = renderMermaidBlocks(in: formatted)
-        let mathResult = renderMathBlocks(in: mermaidResult.html, with: math)
-        return FootnoteDefinitionRenderResult(
-            html: mathResult.html,
-            containsMath: mathResult.containsMath,
-            containsMermaid: mermaidResult.containsMermaid
-        )
-    }
-
-    private static func footnoteDefinitionID(number: Int) -> String {
-        "fn-\(number)"
-    }
-
-    private static func footnoteReferenceID(number: Int, ordinal: Int) -> String {
-        ordinal == 1 ? "fnref-\(number)" : "fnref-\(number)-\(ordinal)"
-    }
-
-    // MARK: - Math (KaTeX)
-
-    private struct MathExtraction {
-        let processedMarkdown: String
-        let blocks: [String]
-        let blockLineCounts: [Int]
-        let inlines: [String]
-    }
-
-    private struct MathRenderResult {
-        let html: String
-        let containsMath: Bool
-    }
-
-    private static let blockMathRegex: NSRegularExpression = {
-        // swiftlint:disable:next force_try
-        try! NSRegularExpression(pattern: #"\$\$([\s\S]+?)\$\$"#)
-    }()
-
-    private static let bracketedBlockMathRegex: NSRegularExpression = {
-        // swiftlint:disable:next force_try
-        try! NSRegularExpression(pattern: #"(?<!\\)\\\[([\s\S]+?)\\\]"#)
-    }()
-
-    // Markdown authors sometimes double the delimiter backslashes so the
-    // Markdown parser emits the single-backslash form expected by MathJax.
-    // Math is extracted before Markdown parsing here, so accept that source
-    // form directly while still rejecting runs of three or more backslashes.
-    private static let markdownEscapedBracketedBlockMathRegex: NSRegularExpression = {
-        // swiftlint:disable:next force_try
-        try! NSRegularExpression(pattern: #"(?<!\\)\\\\\[([\s\S]+?)\\\\\]"#)
-    }()
-
-    // Reject leading `\$` (escaped) and require non-whitespace adjacent to
-    // delimiters so prose like "$5 and $10" doesn't match.
-    private static let inlineMathRegex: NSRegularExpression = {
-        // swiftlint:disable:next force_try
-        try! NSRegularExpression(pattern: #"(?<!\\)\$(?=\S)([^\$\n]+?)(?<=\S)\$"#)
-    }()
-
-    private static let parenthesizedInlineMathRegex: NSRegularExpression = {
-        // swiftlint:disable:next force_try
-        try! NSRegularExpression(pattern: #"(?<!\\)\\\(([^\n]+?)\\\)"#)
-    }()
-
-    private static let markdownEscapedParenthesizedInlineMathRegex: NSRegularExpression = {
-        // swiftlint:disable:next force_try
-        try! NSRegularExpression(pattern: #"(?<!\\)\\\\\(([^\n]+?)\\\\\)"#)
-    }()
-
-    // Fenced code block. Group 1 = backtick run, group 2 = info string, group 3 = body.
-    private static let codeFenceRegex: NSRegularExpression = {
-        // swiftlint:disable:next force_try
-        try! NSRegularExpression(
-            pattern: #"(?m)^(`{3,})[ \t]*([^\n`]*)\n([\s\S]*?)\n\1[ \t]*$"#
-        )
-    }()
-
-    // Inline code span: matched-length backtick runs that are not adjacent to other
-    // backticks. Mirrors CommonMark so spans like `` ` ```math ` `` (single-backtick
-    // delimiters around three inner backticks) tokenize correctly.
-    private static let inlineCodeRegex: NSRegularExpression = {
-        // swiftlint:disable:next force_try
-        try! NSRegularExpression(pattern: #"(?<!`)(`+)(?!`)([^\n]*?)(?<!`)\1(?!`)"#)
-    }()
-
-    // First alternative captures attributes+kind+index for a paragraph-wrapped block token
-    // (the common case after swift-markdown wraps the standalone token); the
-    // second captures a bare token. The wrapper is stripped in either case for
-    // block kind to keep the resulting `<div>` out of an enclosing `<p>`.
-    private static let mathTokenRegex: NSRegularExpression = {
-        // swiftlint:disable:next force_try
-        try! NSRegularExpression(
-            pattern: #"<p\b([^>]*)>MdPreviewMath(Block|Inline)(\d+)Token</p>|MdPreviewMath(Block|Inline)(\d+)Token"#
-        )
-    }()
-
-    private static func extractMath(from markdown: String) -> MathExtraction {
-        var blocks: [String] = []
-        var blockLineCounts: [Int] = []
-        var inlines: [String] = []
-        var protected: [String] = []
-
-        let nsMarkdown = markdown as NSString
-        let fenceMatches = codeFenceRegex.matches(
-            in: markdown,
-            range: NSRange(location: 0, length: nsMarkdown.length)
-        )
-        var afterFences = ""
-        afterFences.reserveCapacity(markdown.count)
-        var fenceCursor = 0
-        for match in fenceMatches {
-            afterFences += nsMarkdown.substring(with: NSRange(
-                location: fenceCursor,
-                length: match.range.location - fenceCursor
-            ))
-            let info = CodeFenceInfo(
-                rawInfoString: nsMarkdown.substring(with: match.range(at: 2))
-            )
-            if info.language == "math" {
-                let body = nsMarkdown.substring(with: match.range(at: 3))
-                blocks.append(body)
-                let fullFence = nsMarkdown.substring(with: match.range)
-                blockLineCounts.append(fullFence.count(where: \.isNewline) + 1)
-                let newlineCount = fullFence.count(where: \.isNewline)
-                // Keep the replacement on the opening line and preserve the
-                // original number of line breaks so later source ranges remain
-                // aligned with the editor buffer.
-                afterFences += "MdPreviewMathBlock\(blocks.count - 1)Token"
-                afterFences += String(repeating: "\n", count: newlineCount)
-            } else {
-                protected.append(nsMarkdown.substring(with: match.range))
-                afterFences += "MdPreviewProtect\(protected.count - 1)Token"
-            }
-            fenceCursor = match.range.location + match.range.length
-        }
-        afterFences += nsMarkdown.substring(from: fenceCursor)
-
-        // Inline code spans next, so $..$ inside `` `$x$` `` is not extracted.
-        let afterInlineCode = replaceFullMatches(of: inlineCodeRegex, in: afterFences) { full in
-            protected.append(full)
-            return "MdPreviewProtect\(protected.count - 1)Token"
-        }
-
-        func extractBlocks(matching regex: NSRegularExpression, from source: String) -> String {
-            let nsSource = source as NSString
-            let matches = regex.matches(
-                in: source,
-                range: NSRange(location: 0, length: nsSource.length)
-            )
-            var result = ""
-            result.reserveCapacity(source.count)
-            var cursor = 0
-            for match in matches {
-                result += nsSource.substring(with: NSRange(
-                    location: cursor,
-                    length: match.range.location - cursor
-                ))
-                let capture = nsSource.substring(with: match.range(at: 1))
-                let fullMatch = nsSource.substring(with: match.range)
-                let newlineCount = fullMatch.count(where: \.isNewline)
-                blockLineCounts.append(newlineCount + 1)
-                result += "MdPreviewMathBlock\(blocks.count)Token"
-                result += String(repeating: "\n", count: newlineCount)
-                blocks.append(capture)
-                cursor = match.range.location + match.range.length
-            }
-            result += nsSource.substring(from: cursor)
-            return result
-        }
-
-        let afterDollarBlockMath = extractBlocks(matching: blockMathRegex, from: afterInlineCode)
-        let afterMarkdownEscapedBlockMath = extractBlocks(
-            matching: markdownEscapedBracketedBlockMathRegex,
-            from: afterDollarBlockMath
-        )
-        let afterBlockMath = extractBlocks(
-            matching: bracketedBlockMathRegex,
-            from: afterMarkdownEscapedBlockMath
-        )
-        let afterDollarInlineMath = replaceMatches(
-            of: inlineMathRegex,
-            in: afterBlockMath
-        ) { capture in
-            defer { inlines.append(capture) }
-            return "MdPreviewMathInline\(inlines.count)Token"
-        }
-        let afterMarkdownEscapedInlineMath = replaceMatches(
-            of: markdownEscapedParenthesizedInlineMathRegex,
-            in: afterDollarInlineMath
-        ) { capture in
-            defer { inlines.append(capture) }
-            return "MdPreviewMathInline\(inlines.count)Token"
-        }
-        let afterInlineMath = replaceMatches(
-            of: parenthesizedInlineMathRegex,
-            in: afterMarkdownEscapedInlineMath
-        ) { capture in
-            defer { inlines.append(capture) }
-            return "MdPreviewMathInline\(inlines.count)Token"
-        }
-
-        let processed = restoreIndexedTokens(
-            in: afterInlineMath,
-            prefix: "MdPreviewProtect",
-            suffix: "Token",
-            replacements: protected
-        )
-
-        return MathExtraction(
-            processedMarkdown: processed,
-            blocks: blocks,
-            blockLineCounts: blockLineCounts,
-            inlines: inlines
-        )
-    }
-
-    private static func renderMathBlocks(in html: String,
-                                         with math: MathExtraction) -> MathRenderResult {
-        guard !math.blocks.isEmpty || !math.inlines.isEmpty else {
-            return MathRenderResult(html: html, containsMath: false)
-        }
-
-        let nsHtml = html as NSString
-        let matches = mathTokenRegex.matches(
-            in: html,
-            range: NSRange(location: 0, length: nsHtml.length)
-        )
-        var rebuilt = ""
-        rebuilt.reserveCapacity(html.count)
-        var cursor = 0
-        for match in matches {
-            rebuilt += nsHtml.substring(with: NSRange(
-                location: cursor,
-                length: match.range.location - cursor
-            ))
-            let wrapped = match.range(at: 2).location != NSNotFound
-            let kindRange = wrapped ? match.range(at: 2) : match.range(at: 4)
-            let indexRange = wrapped ? match.range(at: 3) : match.range(at: 5)
-            let isBlock = nsHtml.substring(with: kindRange) == "Block"
-            let index = Int(nsHtml.substring(with: indexRange)) ?? 0
-            let latex = isBlock ? math.blocks[index] : math.inlines[index]
-            let escaped = htmlEscape(latex)
-            var sourceAttributes = wrapped
-                ? nsHtml.substring(with: match.range(at: 1))
-                : ""
-            if isBlock, index < math.blockLineCounts.count {
-                sourceAttributes = expandSourceEnd(
-                    in: sourceAttributes,
-                    lineCount: math.blockLineCounts[index]
-                )
-            }
-            rebuilt += isBlock
-                ? "<div\(sourceAttributes) class=\"math math-display\">\(escaped)</div>"
-                : "<span class=\"math math-inline\">\(escaped)</span>"
-            cursor = match.range.location + match.range.length
-        }
-        rebuilt += nsHtml.substring(from: cursor)
-        return MathRenderResult(html: rebuilt, containsMath: true)
-    }
-
-    private static func expandSourceEnd(in attributes: String, lineCount: Int) -> String {
-        guard lineCount > 1,
-              let startRange = attributes.range(of: #"data-source-start="\d+""#,
-                                                options: .regularExpression),
-              let start = Int(attributes[startRange].dropFirst(19).dropLast()) else {
-            return attributes
-        }
-        let end = start + lineCount - 1
-        return attributes.replacingOccurrences(
-            of: #"data-source-end="\d+""#,
-            with: "data-source-end=\"\(end)\"",
-            options: .regularExpression
-        )
-    }
-
-    // Debug-only perf instrumentation. Routes labelled timings through the
-    // host bridge so `[mdp-perf +Xms]` entries land in Xcode's console while
-    // diagnosing load-phase regressions. Compiled out of release builds —
-    // no-op shims keep call sites unchanged.
-    #if DEBUG
-    private static let perfBridgeScript = """
-    const perfT0 = (typeof performance !== 'undefined' && performance.now)
-        ? performance.now() : 0;
-    function perfNow() {
-        return (typeof performance !== 'undefined' && performance.now)
-            ? performance.now() - perfT0 : 0;
-    }
-    function perfLog(label, detail) {
-        const dt = perfNow().toFixed(1);
-        const msg = '[mdp-perf +' + dt + 'ms] ' + label
-            + (detail !== undefined ? ' ' + detail : '');
-        try { post({ kind: 'log', message: msg }); } catch (e) {}
-    }
-    window.MdPreviewPerf = { now: perfNow, log: perfLog, t0: perfT0 };
-    perfLog('script eval');
-
-    if (typeof PerformanceObserver === 'function') {
-        try {
-            // Disconnect after FCP — paint emits at most two entries
-            // (first-paint, first-contentful-paint), no need to keep the
-            // observer pinned for the WebView's lifetime.
-            const seen = new Set();
-            const po = new PerformanceObserver((list) => {
-                for (const entry of list.getEntries()) {
-                    perfLog('paint:' + entry.name, entry.startTime.toFixed(1) + 'ms');
-                    seen.add(entry.name);
-                }
-                if (seen.has('first-contentful-paint')) po.disconnect();
-            });
-            po.observe({ type: 'paint', buffered: true });
-        } catch (e) {}
-    }
-    """
-    #else
-    private static let perfBridgeScript = """
-    function perfNow() { return 0; }
-    function perfLog() {}
-    window.MdPreviewPerf = { now: perfNow, log: perfLog };
-    """
-    #endif
-
-    // Always-on host bridge: pushes the document height to the AppKit host via
-    // a WKScriptMessageHandler instead of having the host poll. Quietly no-ops
-    // when the bridge isn't installed (e.g. Quick Look render).
-    // Internal so the WebKit regression tests can exercise the exact
-    // `MdPreview.update` pipeline shipped by the app with the bundled
-    // DOMPurify and morphdom runtimes.
-    static let hostBridgeScript: String = """
-    <script>
-    (() => {
-        const localized = {
-            copy: \(javaScriptStringLiteral(NSLocalizedString("Copy", comment: "Code block copy button"))),
-            copied: \(javaScriptStringLiteral(NSLocalizedString("Copied", comment: "Code block copy confirmation"))),
-            copyCode: \(javaScriptStringLiteral(NSLocalizedString("Copy code", comment: "Code block copy button accessibility label"))),
-            codeCopied: \(javaScriptStringLiteral(NSLocalizedString("Code copied", comment: "Code block copy confirmation accessibility label")))
-        };
-        let hasHostBridge = false;
-        const post = (() => {
-            try {
-                const h = window.webkit && window.webkit.messageHandlers
-                    && window.webkit.messageHandlers.mdPreviewHost;
-                if (!h) return () => false;
-                hasHostBridge = true;
-                return (msg) => {
-                    h.postMessage(msg);
-                    return true;
-                };
-            } catch (e) { return () => false; }
-        })();
-
-        \(perfBridgeScript)
-
-        function measureHeight() {
-            const body = document.body;
-            const article = document.querySelector('.markdown-body');
-            if (!body || !article) return 1;
-            const rect = article.getBoundingClientRect();
-            const cs = getComputedStyle(body);
-            const pt = parseFloat(cs.paddingTop) || 0;
-            const pb = parseFloat(cs.paddingBottom) || 0;
-            return Math.max(rect.bottom + pb, pt + article.scrollHeight + pb, 1);
-        }
-
-        let last = -1;
-        let raf = 0;
-
-        function pushHeight() {
-            if (raf) return;
-            raf = requestAnimationFrame(() => {
-                raf = 0;
-                const h = Math.ceil(measureHeight());
-                if (h !== last) {
-                    last = h;
-                    post({ kind: 'height', value: h });
-                }
-            });
-        }
-
-        // Scroll bridge: with compositor-scrolled WKWebView (macOS 26 SDK)
-        // the host can't observe scrolling natively — the page reports it.
-        let lastScroll = -1;
-        let scrollRaf = 0;
-        function pushScroll() {
-            if (scrollRaf) return;
-            scrollRaf = requestAnimationFrame(() => {
-                scrollRaf = 0;
-                const y = window.scrollY || document.documentElement.scrollTop || 0;
-                if (y !== lastScroll) {
-                    lastScroll = y;
-                    post({ kind: 'scrollPosition', value: y });
-                }
-            });
-        }
-        window.addEventListener('scroll', pushScroll, { passive: true });
-
-        window.MdPreviewHost = { pushHeight, measureHeight };
-
-        function elementForEventTarget(target) {
-            if (target instanceof Element) return target;
-            if (target && target.parentElement instanceof Element) return target.parentElement;
-            return document.activeElement instanceof Element ? document.activeElement : null;
-        }
-
-        function keyBelongsToFocusedControl(target) {
-            const el = elementForEventTarget(target);
-            if (!el) return false;
-            if (el.isContentEditable) return true;
-            return !!el.closest([
-                'button',
-                'input',
-                'select',
-                'textarea',
-                'summary',
-                'audio',
-                'video',
-                '[contenteditable]',
-                '[role="button"]',
-                '[role="checkbox"]',
-                '[role="switch"]',
-                '[role="textbox"]',
-                '[role="combobox"]',
-                '[role="listbox"]',
-                '[role="menuitem"]'
-            ].join(','));
-        }
-
-        function handlePreviewScrollKey(event) {
-            if (event.defaultPrevented || event.metaKey || event.ctrlKey || event.altKey || event.isComposing) return false;
-            if (keyBelongsToFocusedControl(event.target)) return false;
-
-            const isSpace = event.key === ' ' || event.key === 'Spacebar' || event.code === 'Space';
-            if (isSpace) {
-                return post({ kind: 'scroll', value: event.shiftKey ? 'pageUp' : 'pageDown' });
-            }
-
-            if (event.shiftKey) return false;
-            const key = (event.key || '').toLowerCase();
-            if (key === 'j') return post({ kind: 'scroll', value: 'lineDown' });
-            if (key === 'k') return post({ kind: 'scroll', value: 'lineUp' });
-            return false;
-        }
-
-        document.addEventListener('keydown', (event) => {
-            if (handlePreviewScrollKey(event)) {
-                event.preventDefault();
-                event.stopPropagation();
-            }
-        }, true);
-
-        function decorateCodeBlocks(root = document) {
-            root.querySelectorAll('pre > code').forEach((code) => {
-                const pre = code.parentElement;
-                if (!pre || pre.dataset.copyButtonReady === '1') return;
-                pre.dataset.copyButtonReady = '1';
-
-                // Wrap pre in a positioned container so the copy button
-                // stays pinned regardless of horizontal scroll inside pre.
-                const wrap = document.createElement('div');
-                wrap.className = 'md-code-wrap';
-                pre.parentNode.insertBefore(wrap, pre);
-                wrap.appendChild(pre);
-
-                const button = document.createElement('button');
-                button.type = 'button';
-                button.className = 'md-code-copy';
-                button.textContent = localized.copy;
-                button.setAttribute('aria-label', localized.copyCode);
-                wrap.appendChild(button);
-            });
-        }
-
-        function cloneSelectionWithoutCopyButtons(selection) {
-            const fragment = document.createDocumentFragment();
-            for (let i = 0; i < selection.rangeCount; i += 1) {
-                fragment.appendChild(selection.getRangeAt(i).cloneContents());
-            }
-            const buttons = fragment.querySelectorAll('.md-code-copy');
-            if (buttons.length === 0) return null;
-            buttons.forEach((button) => button.remove());
-            return fragment;
-        }
-
-        function plainTextFromFragment(fragment) {
-            const div = document.createElement('div');
-            div.appendChild(fragment.cloneNode(true));
-            return div.innerText || div.textContent || '';
-        }
-
-        function htmlFromFragment(fragment) {
-            const div = document.createElement('div');
-            div.appendChild(fragment.cloneNode(true));
-            return div.innerHTML;
-        }
-
-        async function copyCodeBlock(button) {
-            const wrap = button.parentElement;
-            const code = wrap && wrap.querySelector('pre > code');
-            if (!code) return;
-            const text = code.textContent || '';
-            let copied = false;
-            try {
-                copied = post({ kind: 'copyCode', value: text });
-            } catch (e) {}
-            if (!copied && navigator.clipboard && navigator.clipboard.writeText) {
-                try {
-                    await navigator.clipboard.writeText(text);
-                    copied = true;
-                } catch (e) {}
-            }
-            if (!copied) return;
-            button.textContent = localized.copied;
-            button.setAttribute('aria-label', localized.codeCopied);
-            button.classList.add('is-copied');
-            clearTimeout(button.__mdCopyTimer);
-            button.__mdCopyTimer = setTimeout(() => {
-                button.textContent = localized.copy;
-                button.setAttribute('aria-label', localized.copyCode);
-                button.classList.remove('is-copied');
-            }, 1100);
-        }
-
-        document.addEventListener('click', (event) => {
-            const button = event.target.closest('.md-code-copy');
-            if (!button) return;
-            event.preventDefault();
-            event.stopPropagation();
-            copyCodeBlock(button);
-        });
-
-        function enableTaskCheckboxes() {
-            if (!hasHostBridge) return;
-            document.querySelectorAll('.task-list-item-checkbox').forEach((checkbox) => {
-                checkbox.disabled = false;
-            });
-        }
-
-        let activeTableCell = null;
-        let nextTableContextToken = 1;
-        let pendingTableContextAction = null;
-        let selectedTablePart = null;
-        let tableCellDrag = null;
-        let suppressNextTableClick = false;
-
-        function tableMessage(cell, operation, value, pendingValue, pendingCell) {
-            const table = cell && cell.closest('table[data-source-start][data-source-end]');
-            if (!table || table.dataset.tableSaving === '1') return false;
-            const start = Number(table.dataset.sourceStart);
-            const end = Number(table.dataset.sourceEnd);
-            const row = Number(cell.dataset.tableRow);
-            const column = Number(cell.dataset.tableColumn);
-            if (![start, end, row, column].every(Number.isInteger)) return false;
-            table.dataset.tableSaving = '1';
-            table.closest('.md-table-editor')?.classList.add('is-saving');
-            const message = { kind: 'tableEdit', operation, start, end, row, column };
-            if (typeof value === 'string') message.value = value;
-            if (typeof pendingValue === 'string') {
-                message.pendingValue = pendingValue;
-                message.pendingRow = Number(pendingCell?.dataset.tableRow ?? row);
-                message.pendingColumn = Number(pendingCell?.dataset.tableColumn ?? column);
-            }
-            return post(message);
-        }
-
-        function finishTableCellEdit(save) {
-            const cell = activeTableCell;
-            if (!cell) return;
-            activeTableCell = null;
-            const original = cell.dataset.tableOriginal || '';
-            const value = (cell.innerText || '').replace(/\\n+/g, ' ').trim();
-            cell.contentEditable = 'false';
-            cell.classList.remove('is-editing');
-            if (!save) {
-                cell.innerHTML = cell.__mdOriginalHTML || '';
-                return;
-            }
-            if (value !== original) {
-                tableMessage(cell, 'setCell', value);
-            } else {
-                cell.innerHTML = cell.__mdOriginalHTML || '';
-            }
-        }
-
-        function beginTableCellEdit(cell) {
-            if (!hasHostBridge || !cell) return false;
-            if (cell === activeTableCell) return true;
-            clearTablePartSelection();
-            finishTableCellEdit(true);
-            // Never fall back to rendered text. Without exact source metadata,
-            // editing could silently flatten links, emphasis, code, images, or
-            // other inline Markdown into plain text.
-            if (!cell.hasAttribute('data-table-markdown')) return false;
-            cell.__mdOriginalHTML = cell.innerHTML;
-            cell.dataset.tableOriginal = cell.dataset.tableMarkdown || '';
-            cell.textContent = cell.dataset.tableOriginal;
-            cell.contentEditable = 'plaintext-only';
-            cell.classList.add('is-editing');
-            activeTableCell = cell;
-            cell.focus();
-            const selection = window.getSelection();
-            if (selection) {
-                const range = document.createRange();
-                range.selectNodeContents(cell);
-                range.collapse(false);
-                selection.removeAllRanges();
-                selection.addRange(range);
-            }
-            return true;
-        }
-
-        function clearTablePartSelection() {
-            document.querySelectorAll('.is-table-part-selected').forEach((cell) => {
-                cell.classList.remove(
-                    'is-table-part-selected',
-                    'is-table-selection-top',
-                    'is-table-selection-right',
-                    'is-table-selection-bottom',
-                    'is-table-selection-left'
-                );
-            });
-            selectedTablePart?.editor.classList.remove(
-                'is-table-row-selected',
-                'is-table-column-selected',
-                'is-table-range-selected'
-            );
-            selectedTablePart?.editor.removeAttribute('aria-label');
-            selectedTablePart = null;
-        }
-
-        function applyTableSelection(cell, kind, bounds) {
-            if (!cell) return;
-            finishTableCellEdit(true);
-            clearTablePartSelection();
-            const editor = cell.closest('.md-table-editor');
-            const table = cell.closest('table');
-            if (!editor || !table) return;
-            const items = Array.from(table.querySelectorAll('[data-table-row][data-table-column]'))
-                .filter((item) => {
-                    const row = Number(item.dataset.tableRow);
-                    const column = Number(item.dataset.tableColumn);
-                    return row >= bounds.top && row <= bounds.bottom
-                        && column >= bounds.left && column <= bounds.right;
-                });
-            items.forEach((item) => {
-                item.classList.add('is-table-part-selected');
-                const row = Number(item.dataset.tableRow);
-                const column = Number(item.dataset.tableColumn);
-                if (row === bounds.top) item.classList.add('is-table-selection-top');
-                if (column === bounds.right) item.classList.add('is-table-selection-right');
-                if (row === bounds.bottom) item.classList.add('is-table-selection-bottom');
-                if (column === bounds.left) item.classList.add('is-table-selection-left');
-            });
-            editor.classList.add(
-                kind === 'row'
-                    ? 'is-table-row-selected'
-                    : kind === 'column'
-                        ? 'is-table-column-selected'
-                        : 'is-table-range-selected'
-            );
-            window.getSelection()?.removeAllRanges();
-            selectedTablePart = { cell, kind, editor, bounds };
-            editor.tabIndex = 0;
-            if (kind === 'range') {
-                const rowCount = bounds.bottom - bounds.top + 1;
-                const columnCount = bounds.right - bounds.left + 1;
-                editor.setAttribute(
-                    'aria-label',
-                    `Selected ${rowCount} rows by ${columnCount} columns.`
-                );
-            } else {
-                const row = Number(cell.dataset.tableRow);
-                const column = Number(cell.dataset.tableColumn);
-                const number = kind === 'row' ? row : column + 1;
-                editor.setAttribute(
-                    'aria-label',
-                    `Selected ${kind} ${number}. Press Delete to remove it.`
-                );
-            }
-            editor.focus({ preventScroll: true });
-        }
-
-        function selectTablePart(cell, operation) {
-            if (!cell) return;
-            const table = cell.closest('table');
-            if (!table) return;
-            const row = Number(cell.dataset.tableRow);
-            const column = Number(cell.dataset.tableColumn);
-            const kind = operation === 'selectRow' ? 'row' : 'column';
-            const bounds = kind === 'row'
-                ? { top: row, right: table.rows[0].cells.length - 1, bottom: row, left: 0 }
-                : { top: 0, right: column, bottom: table.rows.length - 1, left: column };
-            applyTableSelection(cell, kind, bounds);
-        }
-
-        function selectTableRange(anchorCell, headCell) {
-            if (!anchorCell || !headCell || anchorCell.closest('table') !== headCell.closest('table')) {
-                return;
-            }
-            const anchorRow = Number(anchorCell.dataset.tableRow);
-            const anchorColumn = Number(anchorCell.dataset.tableColumn);
-            const headRow = Number(headCell.dataset.tableRow);
-            const headColumn = Number(headCell.dataset.tableColumn);
-            applyTableSelection(anchorCell, 'range', {
-                top: Math.min(anchorRow, headRow),
-                right: Math.max(anchorColumn, headColumn),
-                bottom: Math.max(anchorRow, headRow),
-                left: Math.min(anchorColumn, headColumn)
-            });
-        }
-
-        function performTableStructure(cell, operation) {
-            if (!cell) return;
-            const table = cell.closest('table');
-            const editingCell = activeTableCell && activeTableCell.closest('table') === table
-                ? activeTableCell : null;
-            let pendingValue = null;
-            if (editingCell) {
-                const value = (editingCell.innerText || '').replace(/\\n+/g, ' ').trim();
-                if (value !== (editingCell.dataset.tableOriginal || '')) pendingValue = value;
-                else editingCell.innerHTML = editingCell.__mdOriginalHTML || '';
-                activeTableCell = null;
-                editingCell.contentEditable = 'false';
-                editingCell.classList.remove('is-editing');
-            }
-            tableMessage(cell, operation, null, pendingValue, editingCell);
-        }
-
-        function requestNativeTableContextMenu(cell) {
-            const table = cell.closest('table');
-            const row = Number(cell.dataset.tableRow);
-            const columnCount = table?.rows[0]?.cells.length || 1;
-            const token = String(nextTableContextToken++);
-            pendingTableContextAction = { token, cell };
-            post({
-                kind: 'tableContextMenu',
-                token,
-                canInsertRowAbove: row > 0,
-                canDuplicateRow: false,
-                canDeleteRow: row > 0,
-                canDeleteColumn: columnCount > 1,
-                showsDuplicateRow: false
-            });
-        }
-
-        function enableTableEditing(root = document) {
-            if (!hasHostBridge) return;
-            root.querySelectorAll('table[data-source-start][data-source-end]').forEach((table) => {
-                if (table.closest('.md-table-editor')) return;
-                const editor = document.createElement('div');
-                editor.className = 'md-table-editor';
-                table.parentNode.insertBefore(editor, table);
-                const scroll = document.createElement('div');
-                scroll.className = 'md-table-scroll';
-                editor.appendChild(scroll);
-                scroll.appendChild(table);
-                table.querySelectorAll('th[data-table-column]').forEach((cell) => {
-                    const column = Number(cell.dataset.tableColumn);
-                    const placeholder = `Column ${column + 1}`;
-                    cell.dataset.placeholder = placeholder;
-                    // `innerText` forces layout when the table is already in
-                    // the live document. Header emptiness only depends on the
-                    // authored content, so `textContent` is sufficient here.
-                    if (!(cell.textContent || '').trim()) cell.textContent = '';
-                    updateTableHeaderAccessibilityLabel(cell);
-                });
-            });
-        }
-
-        function updateTableHeaderAccessibilityLabel(cell) {
-            const placeholder = cell.dataset.placeholder;
-            if (!placeholder) return;
-            if ((cell.textContent || '').trim()) cell.removeAttribute('aria-label');
-            else cell.setAttribute('aria-label', placeholder);
-        }
-
-        if (hasHostBridge) {
-            // One delegated listener covers both initial and morphed tables.
-            document.addEventListener('input', (event) => {
-                const cell = event.target.closest?.(
-                    '.md-table-editor th[data-table-column]'
-                );
-                if (cell) updateTableHeaderAccessibilityLabel(cell);
-            });
-        }
-
-        document.addEventListener('mousedown', (event) => {
-            if (event.button !== 0) return;
-            const cell = event.target.closest?.('.md-table-editor th, .md-table-editor td');
-            if (!cell) return;
-            const row = Number(cell.dataset.tableRow);
-            const column = Number(cell.dataset.tableColumn);
-            if (!Number.isInteger(row) || row < 0 || !Number.isInteger(column)) return;
-            tableCellDrag = {
-                cell,
-                table: cell.closest('table'),
-                row,
-                column,
-                head: cell,
-                active: false
-            };
-        }, true);
-
-        document.addEventListener('mousemove', (event) => {
-            if (!tableCellDrag) return;
-            const hitTarget = document.elementFromPoint?.(event.clientX, event.clientY);
-            const cell = hitTarget?.closest?.('.md-table-editor th, .md-table-editor td')
-                || event.target.closest?.('.md-table-editor th, .md-table-editor td');
-            if (!cell || cell.closest('table') !== tableCellDrag.table) {
-                return;
-            }
-            if (cell === tableCellDrag.cell && !tableCellDrag.active) return;
-            if (cell === tableCellDrag.head) return;
-            event.preventDefault();
-            tableCellDrag.active = true;
-            tableCellDrag.head = cell;
-            selectTableRange(tableCellDrag.cell, cell);
-        }, true);
-
-        document.addEventListener('mouseup', (event) => {
-            if (tableCellDrag?.active) {
-                event.preventDefault();
-                window.getSelection()?.removeAllRanges();
-                suppressNextTableClick = true;
-            }
-            tableCellDrag = null;
-        }, true);
-
-        document.addEventListener('click', (event) => {
-            if (suppressNextTableClick) {
-                suppressNextTableClick = false;
-                event.preventDefault();
-                event.stopPropagation();
-                return;
-            }
-            const cell = event.target.closest('.md-table-editor th, .md-table-editor td');
-            if (!cell) {
-                clearTablePartSelection();
-                finishTableCellEdit(true);
-                return;
-            }
-            if (event.target.closest('a, button, input')) return;
-            if (beginTableCellEdit(cell)) event.preventDefault();
-        });
-
-        document.addEventListener('contextmenu', (event) => {
-            const cell = event.target.closest('.md-table-editor th, .md-table-editor td');
-            if (!cell) return;
-            event.preventDefault();
-            beginTableCellEdit(cell);
-            requestNativeTableContextMenu(cell);
-        });
-
-        document.addEventListener('keydown', (event) => {
-            if (selectedTablePart && event.target === selectedTablePart.editor) {
-                if (event.key === 'Escape') {
-                    event.preventDefault();
-                    clearTablePartSelection();
-                    return;
-                }
-                if (event.key === 'Backspace' || event.key === 'Delete') {
-                    event.preventDefault();
-                    if (selectedTablePart.kind === 'range') return;
-                    const selection = selectedTablePart;
-                    clearTablePartSelection();
-                    performTableStructure(
-                        selection.cell,
-                        selection.kind === 'row' ? 'deleteRow' : 'deleteColumn'
-                    );
-                    return;
-                }
-            }
-            if (!activeTableCell || event.target !== activeTableCell) return;
-            if (event.key === 'Escape') {
-                event.preventDefault();
-                finishTableCellEdit(false);
-            } else if (event.key === 'Enter' || event.key === 'Tab') {
-                event.preventDefault();
-                finishTableCellEdit(true);
-            }
-        });
-
-        document.addEventListener('paste', (event) => {
-            if (!activeTableCell || event.target !== activeTableCell || !event.clipboardData) return;
-            event.preventDefault();
-            document.execCommand('insertText', false, event.clipboardData.getData('text/plain'));
-        });
-
-        document.addEventListener('change', (event) => {
-            const checkbox = event.target.closest('.task-list-item-checkbox');
-            if (!checkbox || !hasHostBridge) return;
-            const item = checkbox.closest('[data-source-line]');
-            const sourceLine = item && Number(item.dataset.sourceLine);
-            if (!Number.isInteger(sourceLine) || sourceLine < 1) {
-                checkbox.checked = !checkbox.checked;
-                return;
-            }
-            checkbox.disabled = true;
-            if (!post({ kind: 'taskCheckbox', line: sourceLine, checked: checkbox.checked })) {
-                checkbox.checked = !checkbox.checked;
-                checkbox.disabled = false;
-            }
-        });
-
-        document.addEventListener('copy', (event) => {
-            const selection = window.getSelection();
-            if (!selection || selection.rangeCount === 0 || !event.clipboardData) return;
-            const fragment = cloneSelectionWithoutCopyButtons(selection);
-            if (!fragment) return;
-            event.clipboardData.setData('text/plain', plainTextFromFragment(fragment));
-            event.clipboardData.setData('text/html', htmlFromFragment(fragment));
-            event.preventDefault();
-        });
-
-        // Vendor lazy-load helpers. rAF is paused while the WKWebView is
-        // offscreen (e.g. during the launch-time warmup before the window
-        // becomes visible), so afterPaint also falls back to setTimeout(50).
-        window.MdPreviewLazy = {
-            afterPaint(cb) {
-                function tick() {
-                    let fired = false;
-                    function fire(via) {
-                        if (!fired) {
-                            fired = true;
-                            perfLog('afterPaint fire', via);
-                            cb();
-                        }
-                    }
-                    requestAnimationFrame(() => requestAnimationFrame(() => fire('rAF')));
-                    setTimeout(() => fire('timeout'), 50);
-                }
-                if (document.readyState === 'loading') {
-                    document.addEventListener('DOMContentLoaded', tick, { once: true });
-                } else {
-                    tick();
-                }
-            },
-            loadScript(src) {
-                return new Promise((resolve, reject) => {
-                    const tStart = perfNow();
-                    perfLog('script append', src);
-                    const s = document.createElement('script');
-                    s.onload = () => {
-                        perfLog('script onload', src + ' (+' + (perfNow() - tStart).toFixed(1) + 'ms)');
-                        resolve();
-                    };
-                    s.onerror = () => reject(new Error('failed: ' + src));
-                    s.src = src;
-                    document.head.appendChild(s);
-                });
-            },
-            // Wires up a renderer whose vendor JS is loaded after first paint.
-            // - registers a reapplier that gates on `loaded`, so fast-path
-            //   updates don't fire the renderer before its bundle has arrived
-            // - on first paint, fetches `src` (and any `extras` after) and
-            //   calls `run`
-            lazyRenderer({ src, extras, run }) {
-                let loaded = false;
-                if (window.MdPreview && window.MdPreview.registerReapplier) {
-                    window.MdPreview.registerReapplier(() => { if (loaded) run(); });
-                }
-                this.afterPaint(async () => {
-                    try {
-                        await this.loadScript(src);
-                        loaded = true;
-                        run();
-                        if (extras) {
-                            for (const e of extras) this.loadScript(e).catch(() => {});
-                        }
-                    } catch (e) {}
-                });
-            }
-        };
-
-        // DOMPurify config. Closes the raw-HTML XSS path on user markdown
-        // (EscapingHTMLFormatter passes block- and inline-HTML through per
-        // CommonMark). Inline event handlers, <script>, <iframe>, <object>,
-        // <embed>, <base>, <meta>, <link>, <style>, and <form> are dropped;
-        // the `style` attribute is stripped to defeat visual-deception
-        // attacks against the copy button (display:none segments inside
-        // <pre><code> would otherwise survive into clipboard textContent).
-        // <button> stays allowed so the mermaid zoom HUD survives sanitize();
-        // without a parent <form> (forbidden above), `formaction` has nothing
-        // to submit to, and on* handlers are stripped by DOMPurify defaults.
-        //
-        // ALLOWED_URI_REGEXP extends DOMPurify's default safe-URL list with
-        // `md-asset:` so markdown image references that resolve to the
-        // document's base directory (![alt](relative/path.png)) keep working.
-        const SANITIZE_CONFIG = {
-            FORBID_TAGS: ['style', 'form', 'iframe', 'object',
-                          'embed', 'meta', 'link', 'base'],
-            FORBID_ATTR: ['style'],
-            ADD_ATTR: ['target'],
-            ALLOWED_URI_REGEXP: /^(?:(?:(?:f|ht)tps?|mailto|tel|callto|sms|cid|xmpp|matrix|md-asset):|[^a-z]|[a-z+.\\-]+(?:[^a-z+.\\-:]|$))/i
-        };
-        function sanitize(html) {
-            if (typeof html !== 'string') return '';
-            if (typeof DOMPurify === 'undefined' || !DOMPurify.sanitize) {
-                // Fail closed: refuse to render rather than risk shipping
-                // unsanitized HTML into innerHTML. This branch fires only if
-                // the bundled purify.min.js is missing from the app bundle.
-                if (window.console && console.error) {
-                    console.error('[md-preview] DOMPurify not loaded; refusing to render article.');
-                }
-                return '';
-            }
-            return DOMPurify.sanitize(html, SANITIZE_CONFIG);
-        }
-
-        // Incremental-update entry point. Each renderer (KaTeX/Mermaid)
-        // registers an idempotent reapplier that re-processes the current
-        // article. Same-flag re-renders skip the WKWebView reload entirely.
-        const reappliers = [];
-        window.MdPreview = window.MdPreview || {};
-        window.MdPreview.performTableContextAction = (token, operation) => {
-            if (!pendingTableContextAction || pendingTableContextAction.token !== token) return false;
-            const pending = pendingTableContextAction;
-            pendingTableContextAction = null;
-            if (operation === 'selectRow' || operation === 'selectColumn') {
-                selectTablePart(pending.cell, operation);
-            } else {
-                performTableStructure(pending.cell, operation);
-            }
-            return true;
-        };
-        window.MdPreview.registerReapplier = (fn) => {
-            if (typeof fn === 'function') reappliers.push(fn);
-        };
-        function mdHash(s) {
-            let h = 5381;
-            for (let i = 0; i < s.length; i++) h = ((h << 5) + h + s.charCodeAt(i)) >>> 0;
-            return h.toString(36);
-        }
-
-        // One row per expensive block kind: the wrapper class, a key prefix,
-        // where the node carrying __mdSrc and the renderer's done flag lives
-        // inside the wrapper (null = the wrapper itself), and where the
-        // source-position attrs live when not on the wrapper. The keying,
-        // preservation, and attr-sync helpers below all derive from this
-        // table, so a new renderer means one new row — not three new branches.
-        const EXPENSIVE_BLOCKS = [
-            { cls: 'mermaid-figure', kind: 'mm',   inner: '.mermaid',   done: 'mmDone',   attrInner: null  },
-            { cls: 'md-code-wrap',   kind: 'code', inner: 'pre > code', done: 'hljsDone', attrInner: 'pre' },
-            { cls: 'math',           kind: 'math', inner: null,         done: 'mathDone', attrInner: null  }
-        ];
-        const EXPENSIVE_SELECTOR = EXPENSIVE_BLOCKS.map((b) => '.' + b.cls).join(', ');
-        function expensiveKindOf(el) {
-            if (!el.classList) return null;
-            return EXPENSIVE_BLOCKS.find((b) => el.classList.contains(b.cls)) || null;
-        }
-        function expensiveSrcNode(el, info) {
-            return info.inner ? el.querySelector(info.inner) : el;
-        }
-
-        // Content-derived keys so morphdom re-pairs unchanged diagrams/math/
-        // code even when blocks are inserted above them. Live nodes hash the
-        // stashed __mdSrc (renderers replace textContent with their output);
-        // incoming nodes hash textContent — the same Swift emitter produced
-        // both strings, so identical source yields identical keys.
-        function keyExpensiveBlocks(root) {
-            const counts = new Map();
-            root.querySelectorAll(EXPENSIVE_SELECTOR).forEach((el) => {
-                const info = expensiveKindOf(el);
-                const srcNode = expensiveSrcNode(el, info);
-                if (!srcNode) return;
-                const src = srcNode.__mdSrc !== undefined ? srcNode.__mdSrc : srcNode.textContent;
-                // Hash memoized per node — live-tree sources are stable
-                // across updates, so only fresh incoming nodes pay the hash.
-                // Kind-prefixed so a hash collision can never pair blocks of
-                // different types (block math and code wraps are both <div>s).
-                let base = srcNode.__mdKeyBase;
-                if (base === undefined || srcNode.__mdKeySrc !== src) {
-                    base = info.kind + '-' + mdHash(src);
-                    srcNode.__mdKeyBase = base;
-                    srcNode.__mdKeySrc = src;
-                }
-                const n = (counts.get(base) || 0) + 1;
-                counts.set(base, n);
-                el.setAttribute('data-md-key', 'k' + base + ':' + n);
-            });
-        }
-
-        // A preserved subtree keeps its pre-shift source metadata, which
-        // would desync scroll handoff and table edits after lines move.
-        // Copy the incoming line attributes onto the live node.
-        function syncSourceAttrs(fromEl, toEl, info) {
-            let from = fromEl;
-            let to = toEl;
-            if (info.attrInner) {
-                from = fromEl.querySelector(info.attrInner);
-                to = toEl.querySelector(info.attrInner);
-                if (!from || !to) return;
-            }
-            for (const name of ['data-source-line', 'data-source-start', 'data-source-end']) {
-                const value = to.getAttribute(name);
-                if (value === null) from.removeAttribute(name);
-                else from.setAttribute(name, value);
-            }
-        }
-
-        // True when the live subtree holds finished renderer output for the
-        // exact source the incoming node carries — morphdom must then leave
-        // it untouched. Anything else (still rendering, changed source)
-        // morphs normally and the reappliers re-render it.
-        function isRenderedForSource(info, fromEl, toEl) {
-            const live = expensiveSrcNode(fromEl, info);
-            const incoming = expensiveSrcNode(toEl, info);
-            return !!live && !!incoming && live.dataset[info.done] === '1'
-                && live.__mdSrc === incoming.textContent;
-        }
-
-        // Both are stateless, so they're built once instead of per update —
-        // MdPreview.update is the per-keystroke-exit/file-change hot path.
-        const SANITIZE_DOM_CONFIG = Object.assign({}, SANITIZE_CONFIG, { RETURN_DOM_FRAGMENT: true });
-        const MORPH_OPTIONS = {
-            childrenOnly: true,
-            getNodeKey: (node) => node.nodeType === 1
-                ? (node.getAttribute('data-md-key') || node.id || undefined)
-                : undefined,
-            onBeforeElUpdated: (fromEl, toEl) => {
-                if (fromEl.isEqualNode(toEl)) return false;
-                // Skipping is all-or-nothing per subtree: letting morphdom
-                // descend would strip the done markers (incoming nodes lack
-                // them) and force a destructive re-render on the next reapply.
-                const info = expensiveKindOf(fromEl);
-                if (info && isRenderedForSource(info, fromEl, toEl)) {
-                    syncSourceAttrs(fromEl, toEl, info);
-                    return false;
-                }
-                if (fromEl.tagName === 'DETAILS') toEl.toggleAttribute('open', fromEl.open);
-                return true;
-            }
-        };
-
-        // `opts.keepHidden` preserves the warmup opacity so the synthetic
-        // Mermaid pre-render doesn't flash on screen. The host then issues a
-        // second update without the flag once the real document arrives,
-        // which clears the inline style and reveals the article.
-        window.MdPreview.update = (articleHTML, opts) => {
-            // Body swaps can carry a document from a different folder — move
-            // the page <base> first so the incoming content's relative URLs
-            // resolve against the right folder.
-            if (opts && opts.baseHref) {
-                const base = document.querySelector('base');
-                if (base && base.getAttribute('href') !== opts.baseHref) {
-                    base.setAttribute('href', opts.baseHref);
-                }
-            }
-            const article = document.querySelector('.markdown-body');
-            if (!article) return;
-            const tStart = perfNow();
-            finishTableCellEdit(false);
-            clearTablePartSelection();
-            // DOM-diff fast path: morph the live article toward the incoming
-            // HTML so finished Mermaid SVGs, KaTeX output, and highlighted
-            // code survive the update instead of being re-rendered. Skipped
-            // for the first populate (empty article), the warmup article,
-            // and whenever morphdom or DOMPurify is missing; any throw
-            // falls back to the innerHTML swap below.
-            const canMorph = !!articleHTML && article.firstElementChild
-                && article.dataset.warmup !== '1'
-                && typeof morphdom === 'function'
-                && typeof DOMPurify !== 'undefined' && DOMPurify.sanitize;
-            let morphed = false;
-            if (canMorph) {
-                try {
-                    const frag = DOMPurify.sanitize(articleHTML, SANITIZE_DOM_CONFIG);
-                    const next = document.createElement('article');
-                    next.appendChild(frag);
-                    // Pre-shape the incoming tree so the decorators' wrappers
-                    // pair one-to-one with the live DOM during the diff.
-                    decorateCodeBlocks(next);
-                    enableTableEditing(next);
-                    keyExpensiveBlocks(article);
-                    keyExpensiveBlocks(next);
-                    morphdom(article, next, MORPH_OPTIONS);
-                    morphed = true;
-                } catch (e) {
-                    perfLog('morphdom fallback', String(e && e.message || e));
-                }
-            }
-            if (!morphed) {
-                article.innerHTML = sanitize(articleHTML);
-            }
-            if (!opts || !opts.keepHidden) {
-                article.style.opacity = '';
-                article.style.pointerEvents = '';
-                // Revealing means the synthetic warmup content is gone — the
-                // hidden populate keeps the flag so the first real document
-                // still takes the guaranteed innerHTML replace above, and
-                // every update after it may morph.
-                delete article.dataset.warmup;
-            }
-            if (articleHTML) {
-                // The morph path already decorated the incoming tree; only
-                // the innerHTML swap leaves fresh undecorated nodes behind.
-                if (!morphed) {
-                    decorateCodeBlocks();
-                    enableTableEditing();
-                }
-                enableTaskCheckboxes();
-                for (const fn of reappliers) {
-                    try { fn(); } catch (e) { /* one bad apple shouldn't block others */ }
-                }
-            }
-            perfLog('MdPreview.update' + (morphed ? ' (morphdom)' : ''), '(+' + (perfNow() - tStart).toFixed(1) + 'ms)');
-            pushHeight();
-        };
-
-        // Initial-load populator. The article body ships inside an inert
-        // <template> element so the parser never fires inline event handlers
-        // on first paint. Pull it out, sanitize, inject. The template is
-        // removed once consumed.
-        function populateFromTemplate() {
-            const tmpl = document.getElementById('md-article-source');
-            if (!tmpl) return;
-            const article = document.querySelector('.markdown-body');
-            const keepHidden = !!(article && article.dataset.warmup === '1');
-            window.MdPreview.update(tmpl.innerHTML, { keepHidden });
-            tmpl.remove();
-        }
-        // Body-end hook: inline-mode documents call this right after the
-        // template parses, before the vendor bundles, so text paints early.
-        window.MdPreview.populateNow = populateFromTemplate;
-
-        function start() {
-            perfLog('start (DOM ready)');
-            populateFromTemplate();
-            decorateCodeBlocks();
-            pushHeight();
-            try {
-                const ro = new ResizeObserver(pushHeight);
-                ro.observe(document.body);
-                const article = document.querySelector('.markdown-body');
-                if (article) ro.observe(article);
-            } catch (e) {}
-            window.addEventListener('md-preview-mermaid-rendered', pushHeight);
-            window.addEventListener('md-preview-math-rendered', pushHeight);
-            window.addEventListener('load', pushHeight);
-        }
-
-        if (document.readyState === 'loading') {
-            document.addEventListener('DOMContentLoaded', start, { once: true });
-        } else {
-            start();
-        }
-    })();
-    </script>
-    """
-
-    private static let katexFallbackScript = """
-    <script>
-    window.addEventListener('load', () => {
-        document.querySelectorAll('.math').forEach((node) => {
-            node.classList.add('math-error');
-            node.textContent = \(javaScriptStringLiteral(
-                NSLocalizedString(
-                    "KaTeX renderer is unavailable.\n\n",
-                    comment: "Math rendering error"
-                )
-            )) + node.textContent;
-        });
-    });
-    </script>
-    """
-
-    /// JS body of `function renderMath()`. Shared between inline and lazy
-    /// modes — only the surrounding wiring (immediate run vs. deferred-on-load)
-    /// differs.
-    private static let katexRenderMathBody = """
-    function renderMath() {
-        document.querySelectorAll('.math').forEach((el) => {
-            if (el.dataset.mathDone === '1') return;
-            const tex = el.textContent;
-            const display = el.classList.contains('math-display');
-            // Pre-render source, stashed so MdPreview.update can pair
-            // unchanged math with its finished output during DOM diffs.
-            el.__mdSrc = tex;
-            try {
-                katex.render(tex, el, {
-                    displayMode: display,
-                    throwOnError: false,
-                    output: 'htmlAndMathml'
-                });
-                el.dataset.mathDone = '1';
-            } catch (err) {
-                el.classList.add('math-error');
-                el.textContent = String((err && err.message) || err);
-                el.dataset.mathDone = '1';
-            }
-        });
-        window.dispatchEvent(new Event('md-preview-math-rendered'));
-    }
-    """
-
-    /// Bundled vendor JS as an inline `<script>` block, or empty when the
-    /// resource is missing so callers fail soft. The `</script` escape is the
-    /// one correctness-sensitive step of inlining — keep it here, in one place.
-    private static func bundledVendorScriptTag(_ name: String, subdir: String) -> String {
-        guard let js = bundledVendorResource(name, ext: "js", subdir: subdir) else {
-            return ""
-        }
-        let safeJS = js.replacingOccurrences(of: "</script", with: "<\\/script")
-        return "<script>\(safeJS)</script>"
-    }
-
-    /// Inline DOMPurify so the bootstrap can call `DOMPurify.sanitize` before
-    /// the first article ever reaches `innerHTML`. Emitted ahead of the host
-    /// bridge so the sanitizer is defined by the time `MdPreview.update` runs.
-    /// If the vendored file is missing (developer setup error), this is empty
-    /// and the bootstrap's `sanitize()` fails closed — rendering an empty
-    /// article rather than shipping unsanitized HTML.
-    /// Cached: bundle contents are immutable for the process lifetime, and
-    /// `render()` runs on every display, so the disk read happens once.
-    private static let dompurifyBlock = bundledVendorScriptTag("purify.min", subdir: "Vendor/DOMPurify")
-
-    /// Inline morphdom so `MdPreview.update` can DOM-diff fast-path updates
-    /// instead of replacing the whole article subtree — finished Mermaid
-    /// SVGs, KaTeX output, and highlighted code survive updates untouched.
-    /// If the vendored file is missing (SPM tests, Quick Look bundle), this
-    /// is empty and `MdPreview.update` keeps its innerHTML fallback. Cached
-    /// for the same reason as `dompurifyBlock`.
-    private static let morphdomBlock = bundledVendorScriptTag("morphdom.min", subdir: "Vendor/Morphdom")
-
-    private static func katexHead(mode: VendorLoading) -> VendorEmission {
-        guard bundledVendorURL("katex.min", ext: "js", subdir: "Vendor/KaTeX") != nil else {
-            return VendorEmission(head: katexFallbackScript)
-        }
-        let css = bundledVendorResource("katex.min", ext: "css", subdir: "Vendor/KaTeX") ?? ""
-
-        let initScript = """
-        <script>
-        (function() {
-            \(katexRenderMathBody)
-            if (window.MdPreview && window.MdPreview.registerReapplier) {
-                window.MdPreview.registerReapplier(renderMath);
-            }
-            if (document.readyState === 'loading') {
-                document.addEventListener('DOMContentLoaded', renderMath, { once: true });
-            } else {
-                renderMath();
-            }
-        })();
-        </script>
-        """
-
-        switch mode {
-        case .inline:
-            let js = bundledVendorResource("katex.min", ext: "js", subdir: "Vendor/KaTeX") ?? ""
-            let copyTex = bundledVendorResource("copy-tex.min", ext: "js", subdir: "Vendor/KaTeX") ?? ""
-            let safeJS = js.replacingOccurrences(of: "</script", with: "<\\/script")
-            let safeCopyTex = copyTex.replacingOccurrences(of: "</script", with: "<\\/script")
-            return VendorEmission(
-                head: "<style>\(css)</style>",
-                body: """
-                <script>\(safeJS)</script>
-                \(initScript)
-                \(safeCopyTex.isEmpty ? "" : "<script>\(safeCopyTex)</script>")
-                """
-            )
-        case .lazy:
-            // CSS stays inline so layout is stable while KaTeX JS streams in.
-            return VendorEmission(head: """
-            <style>\(css)</style>
-            <script>
-            (function() {
-                \(katexRenderMathBody)
-                window.MdPreviewLazy.lazyRenderer({
-                    src: '\(MarkdownAssetScheme.vendorURL("katex.min.js"))',
-                    extras: ['\(MarkdownAssetScheme.vendorURL("copy-tex.min.js"))'],
-                    run: renderMath,
-                });
-            })();
-            </script>
-            """)
-        }
-    }
-
-    private static func bundledVendorURL(_ name: String,
-                                         ext: String,
-                                         subdir: String) -> URL? {
-        let bundles = [Bundle.main, Bundle(for: MarkdownHTMLBundleToken.self)]
-        for bundle in bundles {
-            if let url = bundle.url(forResource: name, withExtension: ext, subdirectory: subdir) {
-                return url
-            }
-            if let url = bundle.url(forResource: name, withExtension: ext) {
-                return url
-            }
-        }
-        return nil
-    }
-
-    private static func bundledVendorResource(_ name: String,
-                                              ext: String,
-                                              subdir: String) -> String? {
-        bundledVendorURL(name, ext: ext, subdir: subdir).flatMap {
-            try? String(contentsOf: $0, encoding: .utf8)
-        }
-    }
-
-    private static func replaceMatches(of regex: NSRegularExpression,
-                                       in source: String,
-                                       transform: (String) -> String) -> String {
-        rewrite(matchesOf: regex, in: source, captureGroup: 1, transform: transform)
-    }
-
-    private static func replaceFullMatches(of regex: NSRegularExpression,
-                                           in source: String,
-                                           transform: (String) -> String) -> String {
-        rewrite(matchesOf: regex, in: source, captureGroup: 0, transform: transform)
-    }
-
-    private static func rewrite(matchesOf regex: NSRegularExpression,
-                                in source: String,
-                                captureGroup: Int,
-                                transform: (String) -> String) -> String {
-        let nsSource = source as NSString
-        let matches = regex.matches(
-            in: source,
-            range: NSRange(location: 0, length: nsSource.length)
-        )
-        guard !matches.isEmpty else { return source }
-        var result = ""
-        result.reserveCapacity(source.count)
-        var cursor = 0
-        for match in matches {
-            result += nsSource.substring(with: NSRange(
-                location: cursor,
-                length: match.range.location - cursor
-            ))
-            result += transform(nsSource.substring(with: match.range(at: captureGroup)))
-            cursor = match.range.location + match.range.length
-        }
-        result += nsSource.substring(from: cursor)
-        return result
-    }
-
-    /// Replaces generated `prefix + decimal index + suffix` placeholders in a
-    /// single forward pass. Re-running `replacingOccurrences` once per token
-    /// made documents with many code spans, code fences, or footnotes scale
-    /// quadratically because each restoration rescanned the whole document.
-    private static func restoreIndexedTokens(in source: String,
-                                             prefix: String,
-                                             suffix: String,
-                                             replacements: [String]) -> String {
-        guard !replacements.isEmpty, source.contains(prefix) else { return source }
-
-        var result = ""
-        result.reserveCapacity(source.count)
-        var cursor = source.startIndex
-
-        while let tokenStart = source.range(
-            of: prefix,
-            range: cursor..<source.endIndex
-        ) {
-            result.append(contentsOf: source[cursor..<tokenStart.lowerBound])
-
-            var digitsEnd = tokenStart.upperBound
-            while digitsEnd < source.endIndex, source[digitsEnd].isNumber {
-                source.formIndex(after: &digitsEnd)
-            }
-
-            guard digitsEnd > tokenStart.upperBound,
-                  source[digitsEnd...].hasPrefix(suffix),
-                  let index = Int(source[tokenStart.upperBound..<digitsEnd]),
-                  replacements.indices.contains(index) else {
-                // This prefix can appear in authored Markdown. Preserve it and
-                // continue searching after the prefix instead of consuming an
-                // unrelated suffix later in the document.
-                result.append(contentsOf: source[tokenStart.lowerBound..<tokenStart.upperBound])
-                cursor = tokenStart.upperBound
-                continue
-            }
-
-            result.append(replacements[index])
-            cursor = source.index(digitsEnd, offsetBy: suffix.count)
-        }
-
-        result.append(contentsOf: source[cursor...])
-        return result
-    }
-
-    private static func htmlEscape(_ string: String) -> String {
-        var out = ""
-        out.reserveCapacity(string.count)
-        for ch in string {
-            switch ch {
-            case "&": out += "&amp;"
-            case "<": out += "&lt;"
-            case ">": out += "&gt;"
-            case "\"": out += "&quot;"
-            default: out.append(ch)
-            }
-        }
-        return out
-    }
-
-    private static func renderFrontmatter(_ raw: String,
-                                          format: MarkdownFrontmatter.Format,
-                                          sourceEndLine: Int) -> String {
-        let entries = MarkdownFrontmatter.parse(raw, format: format)
-        guard !entries.isEmpty else { return "" }
-
-        let rows = entries.map { entry in
-            let valueHTML: String
-            if let items = entry.items {
-                valueHTML = items.map {
-                    "<span class=\"md-fm-pill\" dir=\"auto\">\(htmlEscape($0))</span>"
-                }.joined()
-            } else if entry.value.isEmpty {
-                valueHTML = "<span class=\"md-fm-empty\" aria-hidden=\"true\"></span>"
-            } else {
-                valueHTML = htmlEscape(entry.value)
-            }
-            return """
-            <tr><th scope="row" dir="auto">\(htmlEscape(entry.key))</th><td dir="auto">\(valueHTML)</td></tr>
-            """
-        }.joined(separator: "\n")
-
-        return """
-        <section class="md-frontmatter" data-source-line="1" data-source-start="1" data-source-end="\(max(1, sourceEndLine))">
-        <table><tbody>
-        \(rows)
-        </tbody></table>
-        </section>
-
-        """
-    }
-
-    // Internal (not private) so regression tests can build JS payloads with
-    // the exact escaping the production bridge uses, like `hostBridgeScript`.
-    static func javaScriptStringLiteral(_ string: String) -> String {
-        guard let data = try? JSONSerialization.data(withJSONObject: [string]),
-              let json = String(data: data, encoding: .utf8) else {
-            return "\"\""
-        }
-        return String(json.dropFirst().dropLast())
-    }
-
-    // MARK: - Code highlighting (highlight.js)
-
-    // Excludes `language-mermaid` since renderMermaidBlocks already lifted
-    // those into `<figure>` containers before this runs.
-    private static let highlightableCodeRegex: NSRegularExpression = {
-        // swiftlint:disable:next force_try
-        try! NSRegularExpression(
-            // Block elements now carry source-line attributes for scroll
-            // handoff, so do not require <pre> and <code> to be bare tags.
-            pattern: #"<pre\b[^>]*>\s*<code\b(?![^>]*\blanguage-mermaid\b)[^>]*>"#
-        )
-    }()
-
-    private static func detectHighlightableCode(in html: String) -> Bool {
-        firstMatch(of: highlightableCodeRegex, in: html) != nil
-    }
-
-    /// Yields via rAF every ~8 ms so the main thread is never pinned for
-    /// more than one frame on docs with many code blocks.
-    /// Internal so the WebKit regression tests can exercise the exact script
-    /// shipped by the app with the bundled highlight.js runtime.
-    static let highlightAllBody = """
-    function autoDetectCodeLanguage(source) {
-        const text = (source || '').trim();
-        if (!text) return '';
-        if (/^#!.*\\b(?:ba)?sh\\b|^\\s*\\$\\s+/.test(text)) return 'bash';
-        if (/\\b(?:resource|data|provider|variable|module)\\s+[\"'][\\w-]+[\"'](?:\\s+[\"'][\\w-]+[\"'])?\\s*\\{|\\bterraform\\s*\\{/.test(text)) return 'hcl';
-        if ((text[0] === '{' && text[text.length - 1] === '}')
-            || (text[0] === '[' && text[text.length - 1] === ']')) {
-            try { JSON.parse(text); return 'json'; } catch (_) {}
-        }
-        if (/^\\s*(?:<!DOCTYPE\\s+html|<html\\b|<(?:div|span|section|article)\\b)/i.test(text)) return 'html';
-        if (/^\\s*(?:async\\s+)?def\\s+\\w+\\s*\\(|^\\s*from\\s+\\w+[\\w.]*\\s+import\\b/m.test(text)) return 'python';
-        if (/\\b(?:import\\s+Foundation|func\\s+\\w+\\s*\\(|@main)\\b|\\b(?:let|var)\\s+\\w+\\s*:\\s*(?:String|Int|Bool|Double|Float)\\b/.test(text)) return 'swift';
-        if (/^\\s*(?:#\\s*include\\s*<iostream>|(?:using\\s+namespace\\s+std|std::\\w+|(?:cout|cin)\\s*(?:<<|>>))\\b)/m.test(text)) return 'cpp';
-        if (/^\\s*(?:#\\s*include\\s*[<"](?:assert|ctype|errno|float|inttypes|limits|math|setjmp|signal|stdarg|stdbool|stddef|stdint|stdio|stdlib|string|time)\\.h[>"]|(?:int|void)\\s+main\\s*\\([^)]*\\)\\s*\\{)/m.test(text)) return 'c';
-        if (/\\b(?:SELECT|INSERT|UPDATE|DELETE|CREATE\\s+(?:TABLE|VIEW|INDEX)|WITH)\\b[\\s\\S]*\\b(?:FROM|INTO|WHERE|AS)\\b/i.test(text)) return 'sql';
-        if (/(?:^|\\n)\\s*(?:[#.]?[A-Za-z][\\w-]*)\\s*\\{[\\s\\S]*:[\\s\\S]*\\}/.test(text)
-            || /@(?:media|keyframes|supports)\\b/.test(text)) return 'css';
-        if (/^\\s*(?:const|let|var)\\s+[A-Za-z_$][\\w$]*\\s*(?::[^=\\n]+)?\\s*=/m.test(text)
-            || /\\b(?:function\\s+\\w+\\s*\\(|console\\.(?:log|error|warn)|=>)/.test(text)) return 'javascript';
-        if (/^\\s*(?:[-A-Za-z_][\\w-]*):\\s*(?:[^:#\\n]|$)/m.test(text)
-            && !/[{};]/.test(text)) return 'yaml';
-
-        // The vendored Terraform grammar has high auto-detection relevance,
-        // so only use highlight.js fallback with a curated language set.
-        const candidates = [
-            'javascript', 'typescript', 'python', 'json', 'css', 'xml',
-            'bash', 'swift', 'go', 'ruby', 'rust', 'c', 'cpp', 'java', 'kotlin',
-            'csharp', 'sql', 'yaml', 'toml'
-        ];
-        const result = hljs.highlightAuto(text, candidates);
-        return result && result.relevance >= 2 ? (result.language || '') : '';
-    }
-
-    function decorateShellOptions(block) {
-        if (!block.classList.contains('language-bash')) return;
-        const textNodes = [];
-        const walker = document.createTreeWalker(block, NodeFilter.SHOW_TEXT);
-        while (walker.nextNode()) {
-            const parent = walker.currentNode.parentElement;
-            if (!parent || parent.closest('.hljs-comment, .hljs-string, .hljs-meta, .hljs-attr')) continue;
-            textNodes.push(walker.currentNode);
-        }
-        const optionPattern = /(^|[\\s=])(-{1,2}[A-Za-z][A-Za-z0-9-]*)(?=$|[=\\s;&|)])/g;
-        textNodes.forEach((node) => {
-            const source = node.nodeValue || '';
-            let match;
-            let cursor = 0;
-            let changed = false;
-            const fragment = document.createDocumentFragment();
-            optionPattern.lastIndex = 0;
-            while ((match = optionPattern.exec(source)) !== null) {
-                const optionStart = match.index + match[1].length;
-                fragment.append(document.createTextNode(source.slice(cursor, optionStart)));
-                const span = document.createElement('span');
-                span.className = 'hljs-attr';
-                span.textContent = match[2];
-                fragment.append(span);
-                cursor = optionStart + match[2].length;
-                changed = true;
-            }
-            if (changed) {
-                fragment.append(document.createTextNode(source.slice(cursor)));
-                node.replaceWith(fragment);
-            }
-        });
-    }
-
-    function highlightAll() {
-        if (typeof hljs === 'undefined') return;
-        if (!document.querySelector('pre > code:not([data-hljs-done="1"])')) return;
-        const blocks = Array.prototype.slice.call(
-            document.querySelectorAll('pre > code:not([data-hljs-done="1"])')
-        );
-        MdPreviewPerf.log('hljs highlightAll start', blocks.length + ' blocks');
-        let i = 0;
-        function step() {
-            const sliceStart = MdPreviewPerf.now();
-            while (i < blocks.length) {
-                const block = blocks[i++];
-                // Pre-render source, stashed so MdPreview.update can pair
-                // unchanged blocks with their highlights during DOM diffs.
-                block.__mdSrc = block.textContent;
-                try {
-                    const explicit = Array.from(block.classList)
-                        .find((name) => name.startsWith('language-'));
-                    if (explicit === 'language-mermaid') {
-                        block.dataset.hljsDone = '1';
-                        continue;
-                    }
-                    if (!explicit) {
-                        const detected = autoDetectCodeLanguage(block.__mdSrc);
-                        if (detected) {
-                            block.classList.add('language-' + detected);
-                            block.dataset.mdDetectedLanguage = detected;
-                        }
-                    }
-                    if (Array.from(block.classList).some((name) => name.startsWith('language-'))) {
-                        hljs.highlightElement(block);
-                        decorateShellOptions(block);
-                    }
-                } catch (e) {
-                    MdPreviewPerf.log('hljs threw', String(e && e.message || e));
-                }
-                block.dataset.hljsDone = '1';
-                if (MdPreviewPerf.now() - sliceStart > 8) break;
-            }
-            if (i < blocks.length) {
-                requestAnimationFrame(step);
-            } else {
-                window.dispatchEvent(new Event('md-preview-hljs-rendered'));
-                MdPreviewPerf.log('hljs all done');
-            }
-        }
-        requestAnimationFrame(step);
-    }
-    """
-
-    /// highlight.js ships light rules plus a system-dark media query. These
-    /// higher-specificity rules let an explicit Quick Look mode override that
-    /// media query without changing automatic mode in the document app.
-    private static let highlightForcedColorSchemeStyles = """
-    html[data-mdp-color-scheme="light"] .hljs { color: #24292e; }
-    html[data-mdp-color-scheme="light"] .hljs-doctag,
-    html[data-mdp-color-scheme="light"] .hljs-keyword,
-    html[data-mdp-color-scheme="light"] .hljs-meta .hljs-keyword,
-    html[data-mdp-color-scheme="light"] .hljs-template-tag,
-    html[data-mdp-color-scheme="light"] .hljs-template-variable,
-    html[data-mdp-color-scheme="light"] .hljs-type,
-    html[data-mdp-color-scheme="light"] .hljs-variable.language_ { color: #d73a49; }
-    html[data-mdp-color-scheme="light"] .hljs-title,
-    html[data-mdp-color-scheme="light"] .hljs-title.class_,
-    html[data-mdp-color-scheme="light"] .hljs-title.class_.inherited__,
-    html[data-mdp-color-scheme="light"] .hljs-title.function { color: #6f42c1; }
-    html[data-mdp-color-scheme="light"] .hljs-attr,
-    html[data-mdp-color-scheme="light"] .hljs-attribute,
-    html[data-mdp-color-scheme="light"] .hljs-literal,
-    html[data-mdp-color-scheme="light"] .hljs-meta,
-    html[data-mdp-color-scheme="light"] .hljs-number,
-    html[data-mdp-color-scheme="light"] .hljs-operator,
-    html[data-mdp-color-scheme="light"] .hljs-selector-attr,
-    html[data-mdp-color-scheme="light"] .hljs-selector-class,
-    html[data-mdp-color-scheme="light"] .hljs-selector-id,
-    html[data-mdp-color-scheme="light"] .hljs-variable { color: #005cc5; }
-    html[data-mdp-color-scheme="light"] .hljs-meta .hljs-string,
-    html[data-mdp-color-scheme="light"] .hljs-regexp,
-    html[data-mdp-color-scheme="light"] .hljs-string { color: #032f62; }
-    html[data-mdp-color-scheme="light"] .hljs-built_in,
-    html[data-mdp-color-scheme="light"] .hljs-symbol { color: #e36209; }
-    html[data-mdp-color-scheme="light"] .hljs-code,
-    html[data-mdp-color-scheme="light"] .hljs-comment,
-    html[data-mdp-color-scheme="light"] .hljs-formula { color: #6a737d; }
-    html[data-mdp-color-scheme="light"] .hljs-name,
-    html[data-mdp-color-scheme="light"] .hljs-quote,
-    html[data-mdp-color-scheme="light"] .hljs-selector-pseudo,
-    html[data-mdp-color-scheme="light"] .hljs-selector-tag { color: #22863a; }
-    html[data-mdp-color-scheme="light"] .hljs-subst,
-    html[data-mdp-color-scheme="light"] .hljs-emphasis,
-    html[data-mdp-color-scheme="light"] .hljs-strong { color: #24292e; }
-    html[data-mdp-color-scheme="light"] .hljs-section { color: #005cc5; }
-    html[data-mdp-color-scheme="light"] .hljs-bullet { color: #735c0f; }
-    html[data-mdp-color-scheme="light"] .hljs-addition { color: #22863a; background-color: #f0fff4; }
-    html[data-mdp-color-scheme="light"] .hljs-deletion { color: #b31d28; background-color: #ffeef0; }
-
-    html[data-mdp-color-scheme="dark"] .hljs { color: #c9d1d9; }
-    html[data-mdp-color-scheme="dark"] .hljs-doctag,
-    html[data-mdp-color-scheme="dark"] .hljs-keyword,
-    html[data-mdp-color-scheme="dark"] .hljs-meta .hljs-keyword,
-    html[data-mdp-color-scheme="dark"] .hljs-template-tag,
-    html[data-mdp-color-scheme="dark"] .hljs-template-variable,
-    html[data-mdp-color-scheme="dark"] .hljs-type,
-    html[data-mdp-color-scheme="dark"] .hljs-variable.language_ { color: #ff7b72; }
-    html[data-mdp-color-scheme="dark"] .hljs-title,
-    html[data-mdp-color-scheme="dark"] .hljs-title.class_,
-    html[data-mdp-color-scheme="dark"] .hljs-title.class_.inherited__,
-    html[data-mdp-color-scheme="dark"] .hljs-title.function { color: #d2a8ff; }
-    html[data-mdp-color-scheme="dark"] .hljs-attr,
-    html[data-mdp-color-scheme="dark"] .hljs-attribute,
-    html[data-mdp-color-scheme="dark"] .hljs-literal,
-    html[data-mdp-color-scheme="dark"] .hljs-meta,
-    html[data-mdp-color-scheme="dark"] .hljs-number,
-    html[data-mdp-color-scheme="dark"] .hljs-operator,
-    html[data-mdp-color-scheme="dark"] .hljs-selector-attr,
-    html[data-mdp-color-scheme="dark"] .hljs-selector-class,
-    html[data-mdp-color-scheme="dark"] .hljs-selector-id,
-    html[data-mdp-color-scheme="dark"] .hljs-variable { color: #79c0ff; }
-    html[data-mdp-color-scheme="dark"] .hljs-meta .hljs-string,
-    html[data-mdp-color-scheme="dark"] .hljs-regexp,
-    html[data-mdp-color-scheme="dark"] .hljs-string { color: #a5d6ff; }
-    html[data-mdp-color-scheme="dark"] .hljs-built_in,
-    html[data-mdp-color-scheme="dark"] .hljs-symbol { color: #ffa657; }
-    html[data-mdp-color-scheme="dark"] .hljs-code,
-    html[data-mdp-color-scheme="dark"] .hljs-comment,
-    html[data-mdp-color-scheme="dark"] .hljs-formula { color: #8b949e; }
-    html[data-mdp-color-scheme="dark"] .hljs-name,
-    html[data-mdp-color-scheme="dark"] .hljs-quote,
-    html[data-mdp-color-scheme="dark"] .hljs-selector-pseudo,
-    html[data-mdp-color-scheme="dark"] .hljs-selector-tag { color: #7ee787; }
-    html[data-mdp-color-scheme="dark"] .hljs-subst,
-    html[data-mdp-color-scheme="dark"] .hljs-emphasis,
-    html[data-mdp-color-scheme="dark"] .hljs-strong { color: #c9d1d9; }
-    html[data-mdp-color-scheme="dark"] .hljs-section { color: #1f6feb; }
-    html[data-mdp-color-scheme="dark"] .hljs-bullet { color: #f2cc60; }
-    html[data-mdp-color-scheme="dark"] .hljs-addition { color: #aff5b4; background-color: #033a16; }
-    html[data-mdp-color-scheme="dark"] .hljs-deletion { color: #ffdcd7; background-color: #67060c; }
-    """
-
-    private static func highlightHead(mode: VendorLoading) -> VendorEmission {
-        guard bundledVendorURL("highlight.min", ext: "js", subdir: "Vendor/Highlight") != nil else {
-            return VendorEmission()
-        }
-        let css = bundledVendorResource("highlight.min", ext: "css", subdir: "Vendor/Highlight") ?? ""
-        let themedCSS = css + "\n" + highlightForcedColorSchemeStyles
-
-        let initScript = """
-        <script>
-        (function() {
-            \(highlightAllBody)
-            if (window.MdPreview && window.MdPreview.registerReapplier) {
-                window.MdPreview.registerReapplier(highlightAll);
-            }
-            if (document.readyState === 'loading') {
-                document.addEventListener('DOMContentLoaded', highlightAll, { once: true });
-            } else {
-                highlightAll();
-            }
-        })();
-        </script>
-        """
-
-        switch mode {
-        case .inline:
-            let js = bundledVendorResource("highlight.min", ext: "js", subdir: "Vendor/Highlight") ?? ""
-            let safeJS = js.replacingOccurrences(of: "</script", with: "<\\/script")
-            return VendorEmission(
-                head: "<style>\(themedCSS)</style>",
-                body: """
-                <script>\(safeJS)</script>
-                \(initScript)
-                """
-            )
-        case .lazy:
-            // CSS stays inline so layout doesn't shift when the JS arrives.
-            return VendorEmission(head: """
-            <style>\(themedCSS)</style>
-            <script>
-            (function() {
-                \(highlightAllBody)
-                window.MdPreviewLazy.lazyRenderer({
-                    src: '\(MarkdownAssetScheme.vendorURL("highlight.min.js"))',
-                    run: highlightAll,
-                });
-            })();
-            </script>
-            """)
-        }
-    }
-
-    // MARK: - Mermaid
-
-    private struct MermaidRenderResult {
-        let html: String
-        let containsMermaid: Bool
-    }
-
-    private static let mermaidRegex: NSRegularExpression = {
-        // swiftlint:disable:next force_try
-        try! NSRegularExpression(
-            pattern: #"<pre\b([^>]*)>\s*<code\b[^>]*class="[^"]*\blanguage-mermaid\b[^"]*"[^>]*>([\s\S]*?)</code>\s*</pre>"#
-        )
-    }()
-
-    private static func renderMermaidBlocks(in html: String) -> MermaidRenderResult {
-        guard html.contains("language-mermaid") else {
-            return MermaidRenderResult(html: html, containsMermaid: false)
-        }
-        let nsHTML = html as NSString
-        let matches = mermaidRegex.matches(
-            in: html,
-            range: NSRange(location: 0, length: nsHTML.length)
-        )
-        guard !matches.isEmpty else {
-            return MermaidRenderResult(html: html, containsMermaid: false)
-        }
-
-        var rendered = ""
-        rendered.reserveCapacity(html.count)
-        var cursor = 0
-        let diagramLabel = htmlEscape(NSLocalizedString("Mermaid diagram", comment: "Mermaid diagram accessibility label"))
-        let zoomOut = htmlEscape(NSLocalizedString("Zoom Out", comment: "Mermaid diagram control"))
-        let resetZoom = htmlEscape(NSLocalizedString("Reset zoom", comment: "Mermaid diagram control"))
-        let zoomIn = htmlEscape(NSLocalizedString("Zoom In", comment: "Mermaid diagram control"))
-        let fillWidth = htmlEscape(NSLocalizedString("Fill width", comment: "Mermaid diagram control"))
-        // The popup is app-only — `MarkdownWebView` compiles its `mermaidPopup`
-        // handler out of the Quick Look extension, so emitting the button there
-        // would leave a control that does nothing when clicked.
-        #if QUICK_LOOK_EXTENSION
-        let popupButton = ""
-        #else
-        let openWindow = htmlEscape(NSLocalizedString("Open in Window", comment: "Mermaid diagram control"))
-        let popupButton = """
-        <button type="button" class="mermaid-hud-btn mermaid-hud-popup" data-mm-act="popup" tabindex="-1" aria-label="\(openWindow)" title="\(openWindow)">⛶</button>
-        """
-        #endif
-        for match in matches {
-            rendered += nsHTML.substring(with: NSRange(
-                location: cursor,
-                length: match.range.location - cursor
-            ))
-            let sourceAttributes = nsHTML.substring(with: match.range(at: 1))
-            let diagram = nsHTML.substring(with: match.range(at: 2))
-            rendered += """
-            <figure\(sourceAttributes) class="mermaid-figure" tabindex="0" role="img" aria-label="\(diagramLabel)">
-            <div class="mermaid-stage"><div class="mermaid">
-            \(diagram)
-            </div></div>
-            <div class="mermaid-hud" aria-hidden="true">
-            <div class="mermaid-hud-group mermaid-hud-zoom">
-            <button type="button" class="mermaid-hud-btn" data-mm-act="out" tabindex="-1" aria-label="\(zoomOut)">−</button>
-            <button type="button" class="mermaid-hud-btn mermaid-hud-level" data-mm-act="reset" tabindex="-1" aria-label="\(resetZoom)">100%</button>
-            <button type="button" class="mermaid-hud-btn" data-mm-act="in" tabindex="-1" aria-label="\(zoomIn)">+</button>
-            </div>
-            <div class="mermaid-hud-group mermaid-hud-actions">
-            <button type="button" class="mermaid-hud-btn mermaid-hud-width" data-mm-act="width" tabindex="-1" aria-label="\(fillWidth)" aria-pressed="false" title="\(fillWidth)"><span class="mermaid-hud-width-symbol" aria-hidden="true">⤢</span></button>
-            \(popupButton)
-            </div>
-            </div>
-            </figure>
-            """
-            cursor = match.range.location + match.range.length
-        }
-        rendered += nsHTML.substring(from: cursor)
-        return MermaidRenderResult(html: rendered, containsMermaid: true)
-    }
-
-    private static let mermaidFallbackScript = """
-    <script>
-    window.addEventListener('load', () => {
-        document.querySelectorAll('.mermaid').forEach((node) => {
-            node.classList.add('mermaid-error');
-            node.textContent = \(javaScriptStringLiteral(
-                NSLocalizedString(
-                    "Mermaid renderer is unavailable.\n\n",
-                    comment: "Mermaid rendering error"
-                )
-            )) + node.textContent;
-        });
-    });
-    </script>
-    """
-
-    /// Mermaid wiring IIFE. Assumes the `mermaid` global has been (or will
-    /// be) defined by the time DOMContentLoaded fires — true for both inline
-    /// vendor `<script>` and `<script defer src=...>` delivery, since `defer`
-    /// scripts run before DOMContentLoaded.
-    /// Mermaid pan/zoom/HUD/popup wiring. Internal so SPM helper tests can
-    /// drive the real bootstrap path (with a stub `mermaid` + message host).
-    static let mermaidInitWiring = """
-    (() => {
-            const fillWidthLabel = \(javaScriptStringLiteral(
-                NSLocalizedString("Fill width", comment: "Mermaid diagram control")
-            ));
-            const fitDiagramLabel = \(javaScriptStringLiteral(
-                NSLocalizedString("Fit diagram", comment: "Mermaid diagram control")
-            ));
-            const states = new WeakMap();
-            const queue = [];
-            let drainPromise = null;
-            let initializedTheme = null;
-            let themeOverride = null;
-
-            function selectedTheme() {
-                const nativeScheme = document.documentElement.dataset.mdpColorScheme;
-                if (nativeScheme) return nativeScheme === 'dark' ? 'dark' : 'default';
-                const dark = window.matchMedia
-                    && window.matchMedia('(prefers-color-scheme: dark)').matches;
-                return dark ? 'dark' : 'default';
-            }
-
-            // Paper printing forces the light palette while mermaid bakes its
-            // theme into the SVG, so the host can pin a theme for the length
-            // of a print operation.
-            function activeTheme() {
-                return themeOverride || selectedTheme();
-            }
-
-            function ensureInit(theme) {
-                if (initializedTheme === theme) return;
-                initializedTheme = theme;
-                mermaid.initialize({
-                    startOnLoad: false,
-                    theme,
-                    securityLevel: 'strict',
-                    fontFamily: '-apple-system, BlinkMacSystemFont, "SF Pro Text", system-ui, sans-serif'
-                });
-            }
-
-            function drain() {
-                if (drainPromise) return drainPromise;
-                drainPromise = (async () => {
-                    while (queue.length) {
-                        const figure = queue.shift();
-                        try {
-                            await renderOne(figure);
-                        } catch (err) {
-                            figure.classList.add('mermaid-error');
-                        }
-                    }
-                })().then(() => {
-                    drainPromise = null;
-                    if (queue.length) return drain();
-                    window.dispatchEvent(new Event('md-preview-mermaid-rendered'));
-                });
-                return drainPromise;
-            }
-
-            // Rendering is viewport-driven, so pagination (print/export) must
-            // force every remaining figure through the queue and wait for the
-            // drain to settle before the page is captured. Passing a theme
-            // pins it (and re-renders mismatched figures) until the next call
-            // without one restores the on-screen theme.
-            async function renderAll(theme) {
-                themeOverride = theme || null;
-                const want = activeTheme();
-                document.querySelectorAll('.mermaid-figure').forEach((figure) => {
-                    const node = figure.querySelector('.mermaid');
-                    if (!node || queue.includes(figure)) return;
-                    if (node.dataset.mmDone !== '1' || node.dataset.mmTheme !== want) {
-                        queue.push(figure);
-                    }
-                });
-                while (queue.length || drainPromise) {
-                    await drain();
-                }
-            }
-
-            async function renderOne(figure) {
-                const theme = activeTheme();
-                ensureInit(theme);
-                const node = figure.querySelector('.mermaid');
-                if (!node) return;
-                if (node.dataset.mmDone === '1') {
-                    if (node.dataset.mmTheme === theme) return;
-                    if (typeof node.__mdSrc !== 'string') return;
-                    // Theme change: put the stashed source back and clear any
-                    // pan/zoom transform now, so a pagination that starts
-                    // before the next frame never captures a stale state.
-                    const prior = states.get(figure);
-                    if (prior && prior.surface) prior.surface.style.transform = '';
-                    node.textContent = node.__mdSrc;
-                    node.removeAttribute('data-processed');
-                    delete node.dataset.mmDone;
-                } else {
-                    // Pre-render source, stashed so MdPreview.update can pair
-                    // unchanged diagrams with their SVGs during DOM diffs.
-                    node.__mdSrc = node.textContent;
-                }
-                try {
-                    await mermaid.run({ nodes: [node], suppressErrors: true });
-                } catch (err) {
-                    figure.classList.add('mermaid-error');
-                    return;
-                }
-                const svg = node.querySelector('svg');
-                if (!svg) {
-                    figure.classList.add('mermaid-error');
-                    return;
-                }
-                node.dataset.mmDone = '1';
-                node.dataset.mmTheme = theme;
-                attachZoom(figure, svg);
-            }
-
-            function attachZoom(figure, svg) {
-                // Normalize sizing: prefer viewBox, drop intrinsic width/height.
-                let vbW, vbH;
-                const vb = svg.viewBox && svg.viewBox.baseVal;
-                if (vb && vb.width && vb.height) {
-                    vbW = vb.width; vbH = vb.height;
-                } else {
-                    vbW = parseFloat(svg.getAttribute('width')) || svg.getBBox().width || 1;
-                    vbH = parseFloat(svg.getAttribute('height')) || svg.getBBox().height || 1;
-                    svg.setAttribute('viewBox', '0 0 ' + vbW + ' ' + vbH);
-                }
-                svg.removeAttribute('width');
-                svg.removeAttribute('height');
-                svg.setAttribute('preserveAspectRatio', 'xMidYMid meet');
-                svg.style.width = '100%';
-                svg.style.height = '100%';
-                const surface = svg.parentElement || svg;
-                surface.style.transformOrigin = '0 0';
-
-                // Stable layout: figure claims height from the diagram's aspect ratio,
-                // capped by max-height so massive diagrams don't push the page.
-                if (vbW > 0 && vbH > 0) {
-                    figure.style.setProperty('--mm-aspect', vbW + ' / ' + vbH);
-                }
-
-                const alreadyWired = states.has(figure);
-                const state = {
-                    tx: 0, ty: 0, scale: 1, min: 1, max: 8,
-                    rect: null, raf: 0, dragging: false,
-                    lastX: 0, lastY: 0, surface,
-                    vbW, vbH, svg
-                };
-                states.set(figure, state);
-                cacheRect(figure);
-
-                // A theme re-render reuses the figure and its listeners; only
-                // the state and HUD need resetting for the fresh SVG.
-                if (alreadyWired) {
-                    apply(figure, state);
-                    return;
-                }
-
-                figure.addEventListener('pointerenter', () => postMermaidHover(true));
-                figure.addEventListener('pointerleave', () => postMermaidHover(false));
-                figure.addEventListener('wheel', onWheel, { passive: false });
-                figure.addEventListener('pointerdown', onPointerDown);
-                figure.addEventListener('dblclick', onDoubleClick);
-                const hud = figure.querySelector('.mermaid-hud');
-                if (hud) hud.addEventListener('click', onHudClick);
-            }
-
-            function cacheRect(figure) {
-                const s = states.get(figure);
-                if (s) s.rect = figure.getBoundingClientRect();
-            }
-
-            function postMermaidHover(value) {
-                try {
-                    window.webkit?.messageHandlers?.mdPreviewHost?.postMessage({
-                        kind: 'mermaidHover',
-                        value
-                    });
-                } catch (_) {}
-            }
-
-            // Text of the closest heading before the figure in document
-            // order, so the popup window can carry some context instead of
-            // a title that's identical for every diagram in the document.
-            function nearestHeadingText(figure) {
-                const headings = document.querySelectorAll('h1, h2, h3, h4, h5, h6');
-                let text = '';
-                for (const h of headings) {
-                    if (h.compareDocumentPosition(figure) & Node.DOCUMENT_POSITION_FOLLOWING) {
-                        text = (h.textContent || '').trim();
-                    } else {
-                        break;
-                    }
-                }
-                return text;
-            }
-
-            // Measure the diagram's natural (viewBox) size and current on-screen
-            // box, then ask the host to open a floating window sized to fit.
-            function openPopup(figure) {
-                const s = states.get(figure);
-                if (!s || !s.svg) return;
-                const svg = s.svg;
-                let naturalW = s.vbW;
-                let naturalH = s.vbH;
-                try {
-                    // getBBox reflects drawn content; prefer it when viewBox
-                    // is missing or clearly wrong.
-                    const bb = svg.getBBox();
-                    if (bb && bb.width > 1 && bb.height > 1) {
-                        if (!(naturalW > 1 && naturalH > 1)) {
-                            naturalW = bb.width;
-                            naturalH = bb.height;
-                        }
-                    }
-                } catch (_) {}
-                if (!s.rect) cacheRect(figure);
-                const r = s.rect || figure.getBoundingClientRect();
-                // Clone without the pan/zoom transform so the popup shows
-                // the pristine rendered diagram. Drive layout purely from
-                // viewBox so the popup's CSS width/height can stretch or
-                // shrink the graphic with the window — done here, before the
-                // SVG leaves the content process, so the popup document
-                // itself needs no script to finish sizing it.
-                const clone = svg.cloneNode(true);
-                clone.removeAttribute('style');
-                clone.setAttribute('preserveAspectRatio', 'xMidYMid meet');
-                clone.removeAttribute('width');
-                clone.removeAttribute('height');
-                clone.style.width = '100%';
-                clone.style.height = '100%';
-                try {
-                    window.webkit?.messageHandlers?.mdPreviewHost?.postMessage({
-                        kind: 'mermaidPopup',
-                        svg: clone.outerHTML,
-                        sectionTitle: nearestHeadingText(figure),
-                        naturalWidth: naturalW,
-                        naturalHeight: naturalH,
-                        displayWidth: r.width,
-                        displayHeight: r.height
-                    });
-                } catch (_) {}
-            }
-
-            function apply(figure, s) {
-                if (s.raf) return;
-                s.raf = requestAnimationFrame(() => {
-                    s.raf = 0;
-                    s.surface.style.transform = 'translate(' + s.tx + 'px,' + s.ty + 'px) scale(' + s.scale + ')';
-                    const lvl = figure.querySelector('.mermaid-hud-level');
-                    if (lvl) lvl.textContent = Math.round(s.scale * 100) + '%';
-                });
-            }
-
-            function zoomAt(figure, x, y, k) {
-                const s = states.get(figure);
-                if (!s) return;
-                const next = Math.max(s.min, Math.min(s.max, s.scale * k));
-                if (next === s.scale) return;
-                const ratio = next / s.scale;
-                s.tx = x - (x - s.tx) * ratio;
-                s.ty = y - (y - s.ty) * ratio;
-                s.scale = next;
-                if (s.scale <= 1.001) { s.tx = 0; s.ty = 0; }
-                apply(figure, s);
-            }
-
-            function reset(figure) {
-                const s = states.get(figure);
-                if (!s) return;
-                s.tx = 0; s.ty = 0; s.scale = 1;
-                apply(figure, s);
-            }
-
-            function toggleWidth(figure) {
-                const expanded = figure.classList.toggle('mermaid-width-expanded');
-                const btn = figure.querySelector('[data-mm-act="width"]');
-                if (btn) {
-                    const label = expanded ? fitDiagramLabel : fillWidthLabel;
-                    btn.textContent = expanded ? '⤡' : '⤢';
-                    btn.setAttribute('aria-label', label);
-                    btn.setAttribute('aria-pressed', String(expanded));
-                    btn.setAttribute('title', label);
-                }
-                reset(figure);
-                requestAnimationFrame(() => {
-                    cacheRect(figure);
-                    window.dispatchEvent(new Event('md-preview-mermaid-rendered'));
-                });
-            }
-
-            function step(figure, factor) {
-                const s = states.get(figure);
-                if (!s) return;
-                if (!s.rect) cacheRect(figure);
-                const r = s.rect;
-                zoomAt(figure, r.width / 2, r.height / 2, factor);
-            }
-
-            function onWheel(e) {
-                // ⌘/Ctrl + wheel zooms; macOS pinch synthesizes wheel + ctrlKey.
-                // Plain wheel falls through to the page scroll (don't preventDefault).
-                if (!(e.ctrlKey || e.metaKey)) return;
-                const figure = e.currentTarget;
-                const s = states.get(figure);
-                if (!s) return;
-                e.preventDefault();
-                if (!s.rect) cacheRect(figure);
-                const r = s.rect;
-                const k = Math.exp(-e.deltaY * 0.01);
-                zoomAt(figure, e.clientX - r.left, e.clientY - r.top, k);
-            }
-
-            function onPointerDown(e) {
-                if (e.button !== 0) return;
-                const figure = e.currentTarget;
-                const s = states.get(figure);
-                if (!s) return;
-                if (e.target.closest('.mermaid-hud')) return;
-                figure.setPointerCapture(e.pointerId);
-                s.dragging = true;
-                s.lastX = e.clientX;
-                s.lastY = e.clientY;
-                figure.addEventListener('pointermove', onPointerMove);
-                figure.addEventListener('pointerup', onPointerUp);
-                figure.addEventListener('pointercancel', onPointerUp);
-            }
-
-            function onPointerMove(e) {
-                const figure = e.currentTarget;
-                const s = states.get(figure);
-                if (!s || !s.dragging) return;
-                s.tx += e.clientX - s.lastX;
-                s.ty += e.clientY - s.lastY;
-                s.lastX = e.clientX;
-                s.lastY = e.clientY;
-                apply(figure, s);
-            }
-
-            function onPointerUp(e) {
-                const figure = e.currentTarget;
-                const s = states.get(figure);
-                if (!s) return;
-                s.dragging = false;
-                figure.removeEventListener('pointermove', onPointerMove);
-                figure.removeEventListener('pointerup', onPointerUp);
-                figure.removeEventListener('pointercancel', onPointerUp);
-            }
-
-            function onDoubleClick(e) {
-                const figure = e.currentTarget;
-                if (e.target.closest('.mermaid-hud')) return;
-                const s = states.get(figure);
-                if (!s) return;
-                if (s.scale > 1.001) {
-                    reset(figure);
-                } else {
-                    if (!s.rect) cacheRect(figure);
-                    const r = s.rect;
-                    zoomAt(figure, e.clientX - r.left, e.clientY - r.top, 2);
-                }
-            }
-
-            function onHudClick(e) {
-                const btn = e.target.closest('[data-mm-act]');
-                if (!btn) return;
-                e.stopPropagation();
-                const figure = btn.closest('.mermaid-figure');
-                if (!figure) return;
-                figure.focus();
-                switch (btn.dataset.mmAct) {
-                    case 'in':    step(figure, 1.25); break;
-                    case 'out':   step(figure, 0.8);  break;
-                    case 'reset': reset(figure);      break;
-                    case 'width': toggleWidth(figure); break;
-                    case 'popup': openPopup(figure);  break;
-                }
-            }
-
-            const ro = new ResizeObserver((entries) => {
-                for (const entry of entries) cacheRect(entry.target);
-            });
-
-            function bootstrap() {
-                const figures = document.querySelectorAll('.mermaid-figure');
-                if (!figures.length) return;
-                const io = new IntersectionObserver((entries) => {
-                    for (const entry of entries) {
-                        if (entry.isIntersecting) {
-                            io.unobserve(entry.target);
-                            queue.push(entry.target);
-                            ro.observe(entry.target);
-                            drain();
-                        }
-                    }
-                }, { rootMargin: '300px 0px' });
-                figures.forEach((f) => io.observe(f));
-            }
-
-            window.MdPreview = window.MdPreview || {};
-            window.MdPreview.mermaidRenderAll = renderAll;
-
-            return { bootstrap };
-        })()
-    """
-
-    private static func mermaidScript(mode: VendorLoading) -> VendorEmission {
-        guard bundledVendorURL("mermaid.min", ext: "js", subdir: "Vendor/Mermaid") != nil else {
-            return VendorEmission(head: mermaidFallbackScript)
-        }
-        switch mode {
-        case .inline:
-            let vendorJS = bundledVendorResource("mermaid.min", ext: "js", subdir: "Vendor/Mermaid") ?? ""
-            let safeVendor = vendorJS.replacingOccurrences(of: "</script", with: "<\\/script")
-            return VendorEmission(
-                body: """
-                <script>
-                \(safeVendor)
-
-                const __mdpMermaid = \(mermaidInitWiring);
-                if (window.MdPreview && window.MdPreview.registerReapplier) {
-                    window.MdPreview.registerReapplier(__mdpMermaid.bootstrap);
-                }
-                if (document.readyState === 'loading') {
-                    document.addEventListener('DOMContentLoaded', __mdpMermaid.bootstrap, { once: true });
-                } else {
-                    __mdpMermaid.bootstrap();
-                }
-                </script>
-                """
-            )
-        case .lazy:
-            return VendorEmission(head: """
-            <script>
-            (() => {
-                let mm = null;
-                window.MdPreviewLazy.lazyRenderer({
-                    src: '\(MarkdownAssetScheme.vendorURL("mermaid.min.js"))',
-                    run: () => {
-                        mm = mm || \(mermaidInitWiring);
-                        mm.bootstrap();
-                    },
-                });
-            })();
-            </script>
-            """)
-        }
-    }
-
-    private final class MarkdownHTMLBundleToken {}
-
-    // Mirrors MarkdownUI's Theme.docC. Top-only margins (bottom: 0), Apple SF
-    // palette (text #1d1d1f / #f5f5f7, link #0066cc / #2997ff, grid #d2d2d7 /
-    // #424245, code bg #f5f5f7 / #2A2828, aside bg #f5f5f7 / #323232), 15px continuous container
-    // radius, horizontal-only table borders.
-    private static let stylesheet = """
-    :root {
-        color-scheme: light dark;
-        --text: #1d1d1f;
-        --secondary: #6e6e73;
-        --link: #0066cc;
-        --aside-bg: #f5f5f7;
-        --aside-border: #696969;
-        --quote-border: #d2d2d7;
-        --code-bg: #f5f5f7;
-        --grid: #d2d2d7;
-    }
-    :root[data-mdp-color-scheme="light"] {
-        color-scheme: light;
-    }
-    :root[data-mdp-color-scheme="dark"] {
-        color-scheme: dark;
-        --text: #f5f5f7;
-        --secondary: #86868b;
-        --link: #2997ff;
-        --aside-bg: #323232;
-        --aside-border: #9a9a9e;
-        --quote-border: #6e6e73;
-        --code-bg: #2A2828;
-        --grid: #424245;
-    }
-    :root[data-mdp-color-scheme],
-    :root[data-mdp-color-scheme] body {
-        background: Canvas;
-    }
-    @media (prefers-color-scheme: dark) {
-        :root:not([data-mdp-color-scheme="light"]) {
-            --text: #f5f5f7;
-            --secondary: #86868b;
-            --link: #2997ff;
-            --aside-bg: #323232;
-            --aside-border: #9a9a9e;
-            --quote-border: #6e6e73;
-            --code-bg: #2A2828;
-            --grid: #424245;
-        }
-    }
-
-    * { box-sizing: border-box; }
-    mark.md-search-highlight {
-        background: #ffd84d;
-        color: #1d1d1f;
-        -webkit-box-decoration-break: clone;
-    }
-    mark.md-search-highlight-current {
-        background: #ffbf00;
-    }
-    .md-search-burst {
-        position: absolute;
-        pointer-events: none;
-        background: rgba(255, 191, 0, 0.5);
-        border-radius: 6px;
-        box-shadow: 0 0 4px rgba(0, 0, 0, 0.12),
-                    0 2px 6px rgba(0, 0, 0, 0.15);
-        z-index: 9999;
-        transform-origin: center center;
-        will-change: transform;
-        animation: md-search-burst 250ms forwards;
-    }
-    /* Per-segment timing: accelerate into the peak (cubic-bezier ease-in),
-       then decelerate out of it (strong ease-out). High matching velocity
-       at the peak means the motion flows through without pausing — the
-       "stuck" feel of multi-stop ease-out keyframes. */
-    @keyframes md-search-burst {
-        0% {
-            transform: scale(1.0);
-            animation-timing-function: cubic-bezier(0.55, 0, 1, 0.45);
-        }
-        50% {
-            transform: scale(1.32);
-            animation-timing-function: cubic-bezier(0, 0.55, 0.45, 1);
-        }
-        100% {
-            transform: scale(1.0);
-        }
-    }
-    @media (prefers-reduced-motion: reduce) {
-        .md-search-burst { animation-duration: 1ms; }
-    }
-    html, body {
-        margin: 0;
-        padding: 0;
-        overflow: hidden;
-    }
-    /* Hide inner scrollers' bars (tables, math) but never match the root:
-       any custom ::-webkit-scrollbar style on <html>/<body> — including a
-       later "restore" override — swaps the page's native macOS overlay
-       scrollbar for WebKit's legacy one. Zero specificity (:where) keeps
-       the pre::-webkit-scrollbar rules below winning for code blocks. */
-    :where(:not(html):not(body))::-webkit-scrollbar {
-        display: none;
-        width: 0;
-        height: 0;
-    }
-    body {
-        font-family: var(--mdp-doc-font, \(bodyFontFamily));
-        font-size: \(bodyFontSize)px;
-        line-height: \(bodyLineHeight);
-        color: var(--text);
-        background: transparent;
-        padding: \(pagePaddingTop)px \(pagePaddingHorizontal)px \(pagePaddingBottom)px;
-        -webkit-font-smoothing: antialiased;
-    }
-
-    article.markdown-body {
-        max-width: \(contentColumnWidth)px;
-        margin-left: auto;
-        margin-right: auto;
-    }
-    article.markdown-body > *:first-child { margin-top: 0 !important; }
-    .md-inline-tab {
-        white-space: pre;
-        tab-size: 4;
-    }
-    .md-source-list-indent-step {
-        display: block;
-        box-sizing: border-box;
-        padding-inline-start: 1.6em;
-    }
-    .md-source-list-line {
-        display: block;
-        margin-top: \(listItemSpacing)px;
-    }
-    .md-source-list-marker {
-        display: inline-block;
-        box-sizing: border-box;
-        width: 1.6em;
-        margin-inline-start: -1.6em;
-        padding-inline-end: 0.45em;
-        text-align: end;
-    }
-    .md-source-task-marker {
-        text-align: center;
-        padding-inline-end: 0.25em;
-    }
-
-    /* Frontmatter properties — a quiet metadata panel. Deliberately
-       quieter than document content: no row borders (content tables own
-       horizontal rules), a muted key column, and a single hairline that
-       hands off to the document body. */
-    .md-frontmatter {
-        margin: 0 0 1.2em;
-        padding: 0 0 1em;
-        border-bottom: 1px solid var(--grid);
-    }
-    .md-frontmatter table {
-        display: table;
-        width: 100%;
-        table-layout: fixed;
-        margin: 0;
-        overflow: visible;
-        font-size: 0.92em;
-        line-height: 1.5;
-    }
-    .md-frontmatter th,
-    .md-frontmatter td {
-        padding: 0.28em 0;
-        border: 0;
-        vertical-align: baseline;
-        overflow-wrap: anywhere;
-        text-align: left;
-    }
-    .md-frontmatter th {
-        width: 26%;
-        padding-right: 1.4em;
-        font-weight: 500;
-        color: var(--secondary);
-    }
-    .md-frontmatter td {
-        white-space: pre-wrap;
-    }
-    .md-fm-pill {
-        display: inline-block;
-        margin: 0 0.4em 0.2em 0;
-        padding: 0.08em 0.7em;
-        border-radius: 999px;
-        background: color-mix(in srgb, var(--link) 12%, transparent);
-        color: var(--link);
-        font-size: 0.95em;
-        overflow-wrap: anywhere;
-    }
-    .md-fm-empty::before {
-        content: "—";
-        color: var(--secondary);
-    }
-
-    p {
-        margin: \(paragraphSpacing)px 0 0;
-    }
-    /* The final blank of a run shrinks to a small gap so a single authored
-       blank plus the next block's margin matches other renderers' paragraph
-       rhythm. Earlier blanks in the run keep their natural line height, so
-       extra authored blanks still grow the gap. */
-    .md-source-blank-line {
-        height: \(blankLineGap)px;
-    }
-    .md-source-blank-line:has(+ .md-source-blank-line) {
-        height: \(sourceLineHeight)px;
-    }
-
-    h1, h2, h3, h4, h5, h6 {
-        font-weight: 600;
-        line-height: 1.18;
-        margin: 1.6em 0 0;
-    }
-    /* Minor-third heading scale: each level steps down visibly, and only
-       the document title carries the heavier weight. */
-    h1 { font-size: 1.802em; font-weight: 700; margin-top: 0.8em; }
-    h2 { font-size: 1.602em; line-height: 1.06; }
-    h3 { font-size: 1.424em; line-height: 1.07; }
-    h4 { font-size: 1.266em; line-height: 1.08; }
-    h5 { font-size: 1.125em; line-height: 1.09; }
-    h6 { font-size: 1em; line-height: 1.24; }
-    /* The blank before a heading shrinks like every final blank; the
-       heading's own margin restores the one-line gap, keeping the total at
-       one source line plus the small breathing room (blank + margin). */
-    .md-source-blank-line + h1,
-    .md-source-blank-line + h2,
-    .md-source-blank-line + h3,
-    .md-source-blank-line + h4,
-    .md-source-blank-line + h5,
-    .md-source-blank-line + h6 {
-        margin-top: \(sourceLineHeight)px;
-    }
-
-    a { color: var(--link); text-decoration: none; }
-    a:hover { text-decoration: underline; }
-    .footnote-ref {
-        font-size: 0.75em;
-        line-height: 0;
-        vertical-align: super;
-    }
-    .footnote-ref a {
-        padding: 0 0.12em;
-    }
-    .footnotes {
-        margin-top: 2.35em;
-        color: var(--text);
-        font-size: 0.9em;
-        line-height: 1.45;
-    }
-    .footnotes hr {
-        margin: 0 0 1em;
-    }
-    .footnotes ol {
-        margin-top: 0;
-        padding-left: 1.45em;
-    }
-    .footnotes li {
-        margin-top: 0.72em;
-        padding-left: 0.12em;
-    }
-    .footnotes li:first-child {
-        margin-top: 0;
-    }
-    .footnotes li > p:first-child {
-        margin-top: 0;
-    }
-    .footnote-backrefs {
-        display: inline-flex;
-        gap: 0.28em;
-        margin-left: 0.28em;
-        white-space: nowrap;
-    }
-    .footnote-backref {
-        font-size: 0.78em;
-        opacity: 0.65;
-        vertical-align: baseline;
-    }
-    .footnote-backref:hover {
-        opacity: 1;
-    }
-
-    code {
-        font-family: \(codeFontFamily);
-        font-size: var(--mdp-code-font-size, 0.88em);
-        padding: 0.18em 0.42em;
-        background: var(--code-bg);
-        border-radius: 6px;
-    }
-    :not(pre) > code {
-        overflow-wrap: anywhere;
-        -webkit-box-decoration-break: clone;
-        box-decoration-break: clone;
-    }
-    pre {
-        position: relative;
-        margin: \(paragraphSpacing)px 0 0;
-        padding: 10px 14px;
-        background: var(--code-bg);
-        border-radius: 15px;
-        overflow-x: auto;
-        line-height: 1.45;
-    }
-    pre::-webkit-scrollbar {
-        display: block;
-        height: 10px;
-        width: 0;
-    }
-    pre::-webkit-scrollbar-track {
-        background: transparent;
-    }
-    pre::-webkit-scrollbar-thumb {
-        background-color: color-mix(in srgb, var(--text) 22%, transparent);
-        border-radius: 10px;
-        border: 3px solid transparent;
-        background-clip: padding-box;
-    }
-    pre:hover::-webkit-scrollbar-thumb {
-        background-color: color-mix(in srgb, var(--text) 38%, transparent);
-    }
-    pre::-webkit-scrollbar-thumb:hover,
-    pre::-webkit-scrollbar-thumb:active {
-        background-color: color-mix(in srgb, var(--text) 55%, transparent);
-    }
-    pre code {
-        /* highlight.js adds display:block with the .hljs class after its
-           deferred pass. Match that layout from first paint so syntax
-           coloring cannot change the code block's line boxes. */
-        display: block;
-        padding: 0;
-        background: transparent;
-        font-size: var(--mdp-code-font-size, 0.88em);
-    }
-    .md-code-wrap {
-        position: relative;
-        margin: \(paragraphSpacing)px 0 0;
-    }
-    .md-code-wrap > pre { margin: 0; }
-    .md-code-copy {
-        position: absolute;
-        top: 8px;
-        right: 8px;
-        appearance: none;
-        min-width: 56px;
-        height: 24px;
-        padding: 0 10px;
-        border: none;
-        border-radius: 8px;
-        color: var(--secondary);
-        background: color-mix(in srgb, var(--text) 10%, var(--code-bg));
-        font: 500 11px/1 -apple-system, BlinkMacSystemFont, "SF Pro Text", system-ui, sans-serif;
-        cursor: pointer;
-        opacity: 0;
-        transition: opacity 120ms ease,
-                    color 120ms ease,
-                    background-color 120ms ease,
-                    transform 120ms ease;
-        user-select: none;
-        -webkit-user-select: none;
-        z-index: 2;
-    }
-    .md-code-wrap:hover .md-code-copy,
-    .md-code-wrap:focus-within .md-code-copy,
-    .md-code-copy.is-copied {
-        opacity: 1;
-    }
-    .md-code-copy:hover {
-        color: var(--text);
-        background: color-mix(in srgb, var(--text) 16%, var(--code-bg));
-    }
-    .md-code-copy:active {
-        background: color-mix(in srgb, var(--text) 22%, var(--code-bg));
-        transform: scale(0.97);
-    }
-    .md-code-copy:focus-visible {
-        outline: none;
-        box-shadow: 0 0 0 3px color-mix(in srgb, AccentColor 60%, transparent);
-    }
-    @media (prefers-reduced-motion: reduce) {
-        .md-code-copy { transition: none; }
-        .md-code-copy:active { transform: none; }
-    }
-    .mermaid-figure {
-        position: relative;
-        margin: \(largeBlockSpacing)px auto 0;
-        background: var(--code-bg);
-        border-radius: 15px;
-        overflow: hidden;
-        outline: none;
-        aspect-ratio: var(--mm-aspect, 4 / 3);
-        max-height: min(70vh, 720px);
-        contain: layout paint;
-    }
-    .mermaid-figure.mermaid-width-expanded {
-        width: 100%;
-        max-height: none;
-    }
-    .mermaid-figure:focus-visible {
-        box-shadow: 0 0 0 3px color-mix(in srgb, AccentColor 60%, transparent);
-    }
-    .mermaid-stage {
-        position: absolute;
-        inset: 0;
-        overflow: hidden;
-        contain: strict;
-    }
-    .mermaid-figure .mermaid-stage { cursor: grab; }
-    .mermaid-figure .mermaid-stage:active { cursor: grabbing; }
-    .mermaid {
-        position: absolute;
-        inset: 0;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        padding: 16px;
-        box-sizing: border-box;
-    }
-    .mermaid svg {
-        display: block;
-        width: 100%;
-        max-width: none !important;
-        height: 100%;
-    }
-    .mermaid-hud {
-        position: absolute;
-        top: 8px;
-        right: 8px;
-        display: flex;
-        flex-wrap: wrap;
-        justify-content: flex-end;
-        gap: 6px 8px;
-        max-width: calc(100% - 16px);
-        opacity: 0;
-        pointer-events: none;
-        transition: opacity 0.12s ease;
-        z-index: 2;
-        font-size: 12px;
-        line-height: 1;
-        color: var(--text);
-    }
-    .mermaid-hud-group {
-        display: flex;
-        gap: 2px;
-        padding: 3px;
-        border-radius: 9px;
-        background: color-mix(in srgb, Canvas 75%, transparent);
-        backdrop-filter: blur(20px) saturate(160%);
-        -webkit-backdrop-filter: blur(20px) saturate(160%);
-        box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12);
-    }
-    .mermaid-figure:hover .mermaid-hud,
-    .mermaid-figure:focus-within .mermaid-hud {
-        opacity: 1;
-        pointer-events: auto;
-    }
-    .mermaid-hud-btn {
-        appearance: none;
-        border: none;
-        background: transparent;
-        color: inherit;
-        font: inherit;
-        font-weight: 500;
-        padding: 5px 9px;
-        border-radius: 6px;
-        cursor: pointer;
-        min-width: 26px;
-        text-align: center;
-    }
-    .mermaid-hud-btn:hover {
-        background: color-mix(in srgb, var(--text) 12%, transparent);
-    }
-    .mermaid-hud-btn:active {
-        background: color-mix(in srgb, var(--text) 18%, transparent);
-    }
-    .mermaid-hud-level {
-        min-width: 46px;
-        font-variant-numeric: tabular-nums;
-    }
-    .mermaid-hud-width {
-        line-height: 12px;
-    }
-    .mermaid-hud-width-symbol {
-        display: inline-block;
-        font-size: 22px;
-        font-weight: 600;
-    }
-    .mermaid-hud-popup {
-        font-size: 16px;
-        line-height: 12px;
-    }
-    @media (prefers-reduced-motion: reduce) {
-        .mermaid-hud { transition: none; }
-    }
-    .mermaid-error {
-        position: static;
-        aspect-ratio: auto;
-        padding: 12px 16px;
-        text-align: left;
-        white-space: pre-wrap;
-        font-family: \(codeFontFamily);
-        font-size: 0.88em;
-    }
-    .math-display {
-        margin: 1.2em 0 0;
-        overflow-x: auto;
-        overflow-y: hidden;
-    }
-    .math-display .katex-display {
-        margin: 0;
-    }
-    .math-error {
-        color: #b00020;
-        background: var(--code-bg);
-        padding: 4px 8px;
-        border-radius: 6px;
-        font-family: \(codeFontFamily);
-        font-size: 0.88em;
-        white-space: pre-wrap;
-    }
-    :root[data-mdp-color-scheme="dark"] .math-error {
-        color: #ff6e6e;
-    }
-    @media (prefers-color-scheme: dark) {
-        :root:not([data-mdp-color-scheme="light"]) .math-error { color: #ff6e6e; }
-    }
-    .katex { direction: ltr !important; unicode-bidi: isolate; }
-
-    blockquote {
-        margin: \(quoteSpacing)px 0 0;
-        padding-inline-start: 1em;
-        border-inline-start: 4px solid var(--quote-border);
-        color: var(--secondary);
-    }
-    blockquote > *:first-child { margin-top: 0; }
-
-    .markdown-alert {
-        margin: \(largeBlockSpacing)px 0 0;
-        padding: 12px 16px;
-        background: var(--aside-bg);
-        border-left: 4px solid var(--aside-border);
-        border-radius: 6px;
-        color: var(--text);
-    }
-    .markdown-alert > *:first-child { margin-top: 0; }
-    .markdown-alert-title {
-        font-weight: 600;
-        margin: 0;
-        display: flex;
-        align-items: center;
-        line-height: 1;
-    }
-    .markdown-alert-icon {
-        width: 1em;
-        height: 1em;
-        margin-right: 0.5em;
-        flex: 0 0 auto;
-        fill: currentColor;
-    }
-    .markdown-alert-note { border-left-color: #0969da; }
-    .markdown-alert-note .markdown-alert-title { color: #0969da; }
-    .markdown-alert-tip { border-left-color: #1a7f37; }
-    .markdown-alert-tip .markdown-alert-title { color: #1a7f37; }
-    .markdown-alert-important { border-left-color: #8250df; }
-    .markdown-alert-important .markdown-alert-title { color: #8250df; }
-    .markdown-alert-warning { border-left-color: #9a6700; }
-    .markdown-alert-warning .markdown-alert-title { color: #9a6700; }
-    .markdown-alert-caution { border-left-color: #d1242f; }
-    .markdown-alert-caution .markdown-alert-title { color: #d1242f; }
-
-    ul, ol { margin: \(paragraphSpacing)px 0 0; padding-left: 1.6em; }
-    ul { list-style-type: "•  "; }
-    /* The text marker stays for copy/paste and for reserving the gutter,
-       but renders transparent; a 0.4em circle is painted in its place.
-       Drawn with a border, not a background, so PDF export keeps it even
-       when backgrounds are not printed. */
-    ul > li::marker { color: transparent; }
-    ul > li { position: relative; }
-    ul > li:not(.task-list-item)::before {
-        content: "";
-        position: absolute;
-        inset-inline-start: -0.9em;
-        top: 0.56em;
-        width: 0;
-        height: 0;
-        border: 0.2em solid var(--text);
-        border-radius: 50%;
-    }
-    li { margin-top: \(listItemSpacing)px; }
-    li:first-child { margin-top: 0; }
-    li > ul, li > ol { margin-top: \(listItemSpacing)px; }
-    li > p:first-child { margin-top: 0; }
-
-    li.task-list-item { list-style: none; }
-    /* Completed tasks read as done — struck through and muted. */
-    li.task-list-item:has(input.task-list-item-checkbox:checked) {
-        color: var(--secondary);
-        text-decoration: line-through;
-    }
-    li.task-list-item > p:first-of-type { display: inline; margin-top: 0; }
-    .task-list-item-checkbox {
-        -webkit-appearance: none;
-        appearance: none;
-        width: 1.55em;
-        height: 1.55em;
-        margin: 0 0.3em 0.1em -1.85em;
-        vertical-align: middle;
-        border: 1.5px solid var(--grid);
-        border-radius: 50%;
-        background: transparent;
-        position: relative;
-        flex: 0 0 auto;
-    }
-    .task-list-item-checkbox:checked {
-        border-color: #007aff;
-        background: #007aff;
-    }
-    .task-list-item-checkbox:not(:disabled) { cursor: pointer; }
-    .task-list-item-checkbox:checked::after {
-        content: "";
-        position: absolute;
-        inset: 0;
-        background-image: url('data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16"><path d="M4.4 8.4 L7 11 L11.6 5.4" fill="none" stroke="white" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/></svg>');
-        background-repeat: no-repeat;
-        background-position: center;
-        background-size: 100% 100%;
-    }
-
-    table {
-        margin: \(largeBlockSpacing)px 0 0;
-        border-collapse: collapse;
-        display: block;
-        overflow-x: auto;
-        max-width: 100%;
-    }
-    th, td {
-        padding: 9px 10px;
-        border-top: 1px solid var(--grid);
-        border-bottom: 1px solid var(--grid);
-        text-align: left;
-    }
-    th { font-weight: 600; }
-
-    .md-table-editor {
-        position: relative;
-        display: inline-block;
-        width: fit-content;
-        margin: \(largeBlockSpacing)px 0 0;
-        max-width: 100%;
-        overflow: visible;
-    }
-    .md-table-scroll {
-        width: fit-content;
-        max-width: 100%;
-        overflow-x: auto;
-    }
-    .md-table-scroll > table { margin-top: 0; }
-    .md-table-editor:focus { outline: none; }
-    .md-table-editor th,
-    .md-table-editor td { cursor: text; }
-    .md-table-editor th[data-placeholder]:empty::before {
-        content: attr(data-placeholder);
-        color: var(--secondary);
-        font-weight: 400;
-        opacity: 0.72;
-        pointer-events: none;
-    }
-    .md-table-editor th.is-editing,
-    .md-table-editor td.is-editing {
-        outline: 2px solid #007aff;
-        outline-offset: -2px;
-        background: color-mix(in srgb, #007aff 8%, transparent);
-        white-space: pre-wrap;
-    }
-    .md-table-editor .is-table-part-selected {
-        --table-selection-top-edge: 0 0 transparent;
-        --table-selection-right-edge: 0 0 transparent;
-        --table-selection-bottom-edge: 0 0 transparent;
-        --table-selection-left-edge: 0 0 transparent;
-        background: color-mix(in srgb, #007aff 14%, Canvas);
-        box-shadow:
-            var(--table-selection-top-edge),
-            var(--table-selection-right-edge),
-            var(--table-selection-bottom-edge),
-            var(--table-selection-left-edge);
-    }
-    .md-table-editor .is-table-selection-top {
-        --table-selection-top-edge: inset 0 1px color-mix(in srgb, #007aff 52%, transparent);
-    }
-    .md-table-editor .is-table-selection-right {
-        --table-selection-right-edge: inset -1px 0 color-mix(in srgb, #007aff 52%, transparent);
-    }
-    .md-table-editor .is-table-selection-bottom {
-        --table-selection-bottom-edge: inset 0 -1px color-mix(in srgb, #007aff 52%, transparent);
-    }
-    .md-table-editor .is-table-selection-left {
-        --table-selection-left-edge: inset 1px 0 color-mix(in srgb, #007aff 52%, transparent);
-    }
-    .md-table-editor.is-saving { opacity: 0.72; }
-
-    hr {
-        border: 0;
-        height: 1px;
-        background: var(--grid);
-        margin: \(hrSpacing)px 0 0;
-    }
-
-    img {
-        display: block;
-        max-width: 100%;
-        margin: 1.6em auto;
-        border-radius: 10px;
-    }
-    /* Keep downscaled images proportional, but let explicit width/height
-       attributes (e.g. GitHub-style <img height="54">) take effect. */
-    img:not([width]):not([height]) {
-        height: auto;
-    }
-    p img {
-        display: inline-block;
-        vertical-align: middle;
-        margin: 0 0.35em 0.35em 0;
-    }
-    p > img:only-child {
-        display: block;
-        margin: 1.6em auto;
-    }
-
-    strong { font-weight: 600; }
-    em { font-style: italic; }
-
-    [dir="rtl"] { text-align: right; }
-
-    /* ---------------------------------------------------------------------
-       Paper-optimized printing.
-
-       WebKit lays print out at a viewport of the printable width in CSS px
-       (96px per inch), and 1 CSS px maps to exactly 0.75pt on paper. Sizing
-       the body in `pt` here therefore lands at that literal point size, with
-       no scaling factor to compensate for. `md-print-size` (injected by the
-       app at print time) overrides the default below.
-
-       The on-screen palette is dark-mode aware; paper is not, so regular
-       printing restores the light values unconditionally. PDF export adds
-       `previewPrintClass` before entering WebKit's print pipeline, which
-       excludes these paper-only changes and preserves the read-only page.
-       --------------------------------------------------------------------- */
-    @media print {
-        :root:not(.\(previewPrintClass)) {
-            color-scheme: light;
-            --text: #1d1d1f;
-            --secondary: #6e6e73;
-            --link: #0066cc;
-            --aside-bg: #f5f5f7;
-            --aside-border: #696969;
-            --quote-border: #d2d2d7;
-            --code-bg: #f5f5f7;
-            --grid: #d2d2d7;
-        }
-        html,
-        body {
-            overflow: visible;
-        }
-        :root:not(.\(previewPrintClass)),
-        :root:not(.\(previewPrintClass)) body {
-            background: #fff;
-        }
-        @page {
-            margin: \(printPageMarginTop) \(printPageMarginSide) \(printPageMarginBottom);
-        }
-        body {
-            -webkit-print-color-adjust: exact;
-            print-color-adjust: exact;
-        }
-        :root:not(.\(previewPrintClass)) body {
-            font-size: \(defaultPrintPointSize)pt;
-            padding: 0;
-        }
-        /* NSPrintInfo owns the page margins, and the print viewport is
-           narrower than the on-screen measure, so the column just fills it. */
-        :root:not(.\(previewPrintClass)) article.markdown-body {
-            max-width: none;
-            margin: 0;
-        }
-
-        /* Interaction affordances are screen-only. */
-        .md-code-copy,
-        .md-search-burst,
-        .mermaid-hud { display: none !important; }
-        mark.md-search-highlight,
-        mark.md-search-highlight-current {
-            background: transparent;
-            color: inherit;
-        }
-
-        /* Splitting these across a page break loses the reading order. */
-        :root:not(.\(previewPrintClass)) pre,
-        :root:not(.\(previewPrintClass)) blockquote,
-        :root:not(.\(previewPrintClass)) table,
-        :root:not(.\(previewPrintClass)) figure,
-        :root:not(.\(previewPrintClass)) .md-frontmatter,
-        :root:not(.\(previewPrintClass)) .markdown-alert {
-            break-inside: avoid;
-        }
-        :root:not(.\(previewPrintClass)) tr,
-        :root:not(.\(previewPrintClass)) li { break-inside: avoid; }
-        :root:not(.\(previewPrintClass)) h1,
-        :root:not(.\(previewPrintClass)) h2,
-        :root:not(.\(previewPrintClass)) h3,
-        :root:not(.\(previewPrintClass)) h4,
-        :root:not(.\(previewPrintClass)) h5,
-        :root:not(.\(previewPrintClass)) h6 { break-after: avoid; }
-        :root:not(.\(previewPrintClass)) pre {
-            white-space: pre-wrap;
-            word-wrap: break-word;
-        }
-        /* Inner scrollers can't scroll on paper — let them wrap instead of
-           clipping their overflow. */
-        :root:not(.\(previewPrintClass)) .md-code-wrap,
-        :root:not(.\(previewPrintClass)) .md-table-scroll,
-        :root:not(.\(previewPrintClass)) table,
-        :root:not(.\(previewPrintClass)) pre {
-            overflow: visible !important;
-        }
-        :root:not(.\(previewPrintClass)) img,
-        :root:not(.\(previewPrintClass)) svg {
-            max-width: 100% !important;
-            height: auto;
-        }
-        /* Anything wider than the printable area makes WebKit shrink the whole
-           document to fit, which silently overrides the chosen point size — a
-           request for 18pt came out at ~15pt. Keep every block inside the
-           measure so the size stays honest. */
-        :root:not(.\(previewPrintClass)) body { overflow-wrap: break-word; }
-        :root:not(.\(previewPrintClass)) table { width: 100%; }
-        :root:not(.\(previewPrintClass)) th,
-        :root:not(.\(previewPrintClass)) td { overflow-wrap: anywhere; }
-        :root:not(.\(previewPrintClass)) pre,
-        :root:not(.\(previewPrintClass)) code { overflow-wrap: anywhere; }
-    }
-
-    """
 }
