+++ tests/swift-tests/Sources/MarkdownHelpers/MarkdownHTML+Footnotes.swift
@@ -1,0 +1,1 @@
+../../../../md-preview/Rendering/MarkdownHTML+Footnotes.swift
