+++ tests/swift-tests/Sources/MarkdownHelpers/MarkdownHTML+Highlight.swift
@@ -1,0 +1,1 @@
+../../../../md-preview/Rendering/MarkdownHTML+Highlight.swift
