+++ tests/swift-tests/Sources/MarkdownHelpers/MarkdownHTML+HostBridge.swift
@@ -1,0 +1,1 @@
+../../../../md-preview/Rendering/MarkdownHTML+HostBridge.swift
