+++ tests/swift-tests/Sources/MarkdownHelpers/MarkdownHTML+KaTeX.swift
@@ -1,0 +1,1 @@
+../../../../md-preview/Rendering/MarkdownHTML+KaTeX.swift
