+++ tests/swift-tests/Sources/MarkdownHelpers/MarkdownHTML+Math.swift
@@ -1,0 +1,1 @@
+../../../../md-preview/Rendering/MarkdownHTML+Math.swift
