+++ tests/swift-tests/Sources/MarkdownHelpers/MarkdownHTML+Mermaid.swift
@@ -1,0 +1,1 @@
+../../../../md-preview/Rendering/MarkdownHTML+Mermaid.swift
